[roll] Roll fuchsia [elfldltl] Use AbiPtr et al in Soname, SymbolInfo, InitFiniInfo
These data structures will be used in the new dynamic linker's
"passive" ABI, so define them to be compatible with the out of
process case.  They must be templated on not only the Elf layout
but on the AbiTraits to use in their pointer and span members.
This also includes their subordinate data structures such as
CompatHash and GnuHash.

These classes are still only fully usable in their LocalAbiTraits
instantiations.  The testing of other instantiations is limited
to verifying that the types can be instantiated with the stub
RemoteAbiTraits and still support default- and copy-construction
and copy-assignment.

Original-Bug: 121817
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/866118
Original-Revision: 58e38a99560e93607aef2b9f6ca55b92ac05aad5
GitOrigin-RevId: 34926703b2f2b9e7be90e4ed7b31496a53cc2fb6
Change-Id: I1d5ca8f52371555dcedad7374508efb47079b155