[roll] Roll fuchsia [riscv] Fix gn gen for minimal.riscv64
We are using zedboot_qemu because we _have_ to have a ZBI, but we do not
want a vbmeta. Even though we are building a zedboot, we can ignore it.
It is not currently possible to _not_ build a recovery slot.

Original-Bug: 127665
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868016
Original-Revision: 864e7d01dde9193025cfb5b3e0f8224094569d5e
GitOrigin-RevId: 2630951feff53344d80a3977098438f00aa211c2
Change-Id: I712c80119b0bb5cef734b756ddbc2a192b4be77e