[roll] Roll fuchsia [netlink] Only send Ack or Done message if needed
Previously, we responded to all GET requests with a Done message and
all non-GET requests with an Ack message. This change updates the
library to only send Done if the GET request was a dump (had the
NLM_F_DUMP flag set), and only send an Ack if the request requested
one (had the NLM_F_ACK flag set).

As a result of this change, some starnix gvisor rtnetlink tests had
to be disabled as they blocked forever. They blocked forever because
the tests would follow the pattern of:
  1) Send request without `NLM_F_ACK` set (don't send Ack).
  2) Read response from message.

Before this change, the implementation would unconditionally respond to
the tests' requests with an Ack and the test would read that message and
fail as it was not what the test expected. With this change, the tests
that do not request an Ack get no response as the requested operations
are not yet supported. This results in the read calls to the (blocking)
rtnetlink socket to never return.

Original-Bug: b/283827094
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/867958
Original-Revision: 169674e5130383c18934dfc4179cec6eefd112ec
GitOrigin-RevId: 4c51c9bc6fb1dcbe09e8bc59c735e5d7e7f73f0b
Change-Id: I27984651616cf82a44190c459e56c07b562d9a02