[roll] Roll fuchsia [profiler] Added process watcher
This cl implements a process watcher for the cpu profiler. It enables it
to monitor a thread and be notified when a monitored process starts a
new thread so that we can monitor it as well.

Since we're now passing around an async dispatcher anyways, this cl
also moves away from sampling on a std::thread and instead samples using
the async loop. This simplifies some locking, since we'll now never access
the data members concurrently and I did not observe a performance
difference in sample overhead.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862145
Original-Revision: b0536e1436abdf0ffb3ab4c2d2ed9a1fdc703cb0
GitOrigin-RevId: 7e5810f5c11137c019fcb5dd1477f3d49cab99bc
Change-Id: I1eee278cd2136f0cecfbfab5e5eecf72b350d1b7