[roll] Roll fuchsia [wlan] Introduce the `event` module and the `Handler` trait.
This change introduces the `event` module and its core `Handler` trait
in hardware simulator.

This is the first in a series of changes that implement a new event
handler framework for writing hardware simulator tests. The `Handler`
trait, its combinators, and its adapters are the primary mechanism that
powers this API. These additions are based on a series of changes that
include a complete port of tests. Some documentation refers to these
subsequent changes.

Test: Unit tests and hardware simulator tests.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/811381
Original-Revision: 262e96882b3469aa1852ea14abbf8b4da081b46b
GitOrigin-RevId: a2a99101280b35e911662fa4d8c2860e41d58902
Change-Id: I4ac029f1d871a50aa99d7540e2e496c4e133c81f