[roll] Roll fuchsia [amlogic-display] Store image pointers instead of handles / IDs.
This change is similar to fxrev.dev/867723 but it's for amlogic-display
driver. `capture_active_id_` is stored as handle types (uint64_t)
which hides the fact that it is actually a raw pointer to ImageInfo.

Instead of storing a handle, we should store a raw pointer to
ImageInfo and convert them to / from uint64_t handles only at the
banjo API boundary.

Test: No-op. amlogic-display-test
Original-Bug: 128704
Original-Bug: 128653
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868378
Original-Revision: ac4f26ef1523b99db34454e970cd2c9b942bc4c8
GitOrigin-RevId: cf0aadcd08d72b1781ac4b621aafb625e1e03e86
Change-Id: I913d0f8e8d0f36f655b4be328f728bc317e1b84b