[roll] Roll fuchsia [bazel_sdk] Fix ArgumentScope on python 3.11
With python 3.11, the Enum class implementation was changed in a way
that breaks the usage in ArgumentScope. The `EXPLICIT` value is
converted from an empty  tuple to an object in a way that causes the `in`
operator on it to fail. We need to mixin `tuple` to the ArgumentScope
type to ensure that the enum values can act as tuples.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/866524
Original-Revision: 828825a6fb529b2987f3db649b70311334d86d12
GitOrigin-RevId: fe5f9ebc2712272aca1c03ddbea823b919ea7ece
Change-Id: I08449322ed70790b6871d7a0c702afeff95246de