[roll] Roll fuchsia [concurrent] Make the SeqLock APIs easier to hold correctly.
Change the SeqLock APIs exposed by libconcurrent to be easier to use
correctly. In particular, make the memory-ordering synchronization
approach (either acq/rel or fence) a part of the lock type itself.
Then, introduce a new payload wrapper type (`SeqLockPayload<T,
LockType>`) which is an thin layer on top of a
`WellDefinedCopyable<T>` wrapper.  The new wrapper type demands that
the user pass the type of the lock the wrapper is meant to work with,
and exposes `Read/Update` methods which do not give the user any
control over the type of synchronization implemented.  Instead, they
simply "do the right thing" based on properties of the lock type.

Additionally:
1) Update the lockdep wrappers present in `kconcurrent` in order to
   work with the new type-linked approach.
2) Update the existing users of SeqLock in the kernel.
3) Update the documentation in `libconcurrent`'s readme to demonstrate
   the new (harder to mess up) way to hold a SeqLock in the kernel.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/810594
Original-Revision: 9a82251106fc682536070fa391f12e79a6a0d3a3
GitOrigin-RevId: 840c624a3661bbb018963a6b7c223a071f1e3094
Change-Id: I4eecf5d06682c5cd15c626f99b009ab44a4eee5a