[roll] Roll fuchsia [audio] Remove is_input from the composite drivers API
Removed in favor of the signal processing topology describing the one
or many input and output pipelines that could not be expressed with
this one field for the whole driver. If the driver exposes only one
input or output that is also described by the signal processing
topology.

Original-Bug: 124865

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/866686
Original-Revision: f5d3ca101e732549986fee3722a65498566fa3ea
GitOrigin-RevId: aa170180ca20f7d8fb5c943c5c8cf8f8a93c7abf
Change-Id: Iff9158493056fe19eac9109ededf71e1f345ac21