[roll] Roll infra/recipes [cleanup] Move RunSteps to the top of every recipe file
As suggested by http://go/fuchsia-recipe-docs#function-ordering, it's
preferred for `RunSteps` to be the first function in every recipe
because it's the entrypoint, and because `GenTests` is suggested to be
the last function, it can be hard to locate `RunSteps` when it's the
second-to-last function as was the case in all of these recipes.

This unblocks us from enforcing the function ordering with shac.

The main downside is that we lose `git blame` history for the affected
lines, but we can work around that by adding this commit's final hash to
the `.git-blame-ignore-revs` file in a follow-up change, which will
exclude this commit from being considered by `git blame` if `git config
blame.ignoreRevsFile .git-blame-ignore-revs` is set.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/868357
Presubmit-Verified: CQ Bot <fuchsia-internal-scoped@luci-project-accounts.iam.gserviceaccount.com>
Original-Revision: d7bb1c93166487193756186c801a946bef625cf0
GitOrigin-RevId: a4534dab6aec7c88c9fe00ef71b383672b3f8f70
Change-Id: I833ecb57fb152211423d0564410000456cfb3d9d