[roll] Roll fuchsia [boot-shim] Introduce DevicetreeBootShim
A DevicetreeItem is a BootShim item where |GetInitMatcher| would return
an instance of a matcher. After this call, it is expected that items are
able to return the number of bytes required to store the zbi item
generated.

The DevicetreeBootShim provides InitDevicetreeItems such that any
DevicetreeItem::GetInitMatcher instance is part of the devictree::Match
called during |IniteDevicetreeItems|.

Test: fx test boot-shim-unittests
Original-Bug: 78675
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831538
Original-Revision: df8917692072a8cdfcd6a8785bb190082bc9a7bd
GitOrigin-RevId: 8b8e37170ebee2ceba7d97de6221e49579209751
Change-Id: I7bc6c4cd70d3e6f247302a2d42e85f0020a7dfab