[roll] Roll fuchsia [concurrent] Rename inc -> inc.h
libconcurrent contains some files which are technically header files,
but are meant to be used in a single specific translation unit in the
environment that libconcurrent is being used.

Generally speaking, these files contain the implementation of a class
which is templated on an adapter class specific for the environment
(the OSAL).  They are meant to be include in a single translation
unit, where the implementation can be manually expanded into the
actual implementations of the functions with the OSAL included.

Until now, these files have ended in a `.inc` extension in an attempt
to distinguish them from normal header files.  Unfortunately, using a
`.inc` extension can prevent things like auto-formatter utilities, and
editor syntax highlighting from working correctly.  So, rename these
files from `*.inc -> *.inc.h`, hopefully preserving the indication
that "This is not a normal header", while restoring the formatting and
syntax highlighting behaviors.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868881
Original-Revision: a12009df640033d2edf2e2f9afaf43f088351c9d
GitOrigin-RevId: e79fb8e37cb3acc534a34f0fc059f48bc008654f
Change-Id: Ic763fbf5f6cc4cf17af6db756befe4541a0cc301