[roll] Roll fuchsia [kernel][jtrace] Add protection against nullptr records.
When JTRACE's "last-per-cpu-event" feature is enabled, it introduces a
rare edge case which can cause a double panic.  If we encounter a
panic, and CPU X has _never_ recorded any events, the record for that
CPU is full of nothing but zeros.

If we attempt to print this record, it means that pointers to what
should be valid string literals (such as the tag), or the
file-function-line record, will be nullptr.  Attempts to render this
information will trigger a panic, usually when we are already
attempting to dump the trace data during a panic.

Add protection against this behavior by always checking for nullptr in
the SafeString helper class, as well a explicitly checking the FFL
record during a dump.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868882
Original-Revision: fc1e987e4c2d101cea57edf895b5918680397413
GitOrigin-RevId: d951d98de9869daed89f31e3e662f960caf940bc
Change-Id: I064252dcb82bcdf909f59b2d851874bdb7224130