[roll] Roll fuchsia [fake-display] Replace display_lock_ with other locks / atomics.
There are a few issues related to usages of display_lock_ in
FakeDisplay:

(1) `current_image_to_capture_` is a pointer to image in
    `imported_images_`. Instead of `display_lock_`,
    `image_lock_` which locks `imported_images_` should be used to
    lock any reference / pointer to the image.

(2) Update (increment) of `current_config_stamp_` doesn't need to
    be within the same critical section as updating the pointer to
    the image to capture.

(3) CheckConfiguration() doesn't need the lock to be held.

(4) SendVsync() used by the test holds display_lock_, while it
    calls the interface OnDisplayVsync() callback, and the
    coordinator may eventually call other DisplayControllerImpl*()
    methods, causing the display_lock_ to be relocked.

As a result, this change removes display_lock_ and replaces usages
of display_lock_ with other locks or atomics: for interface access
interface_lock_ is used; for current image access image_lock_ is
used; atomic (with relaxed memory order) is used to access
the currently applied config stamp which is sufficient to guarantee
all changes to the images to capture will be visible to the capture
thread after the Vsync gets handled by the coordinator.

Test: no behavior change expected.
Original-Bug: 128578
Multiply: display-coordinator-unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868181
Original-Revision: 1617750165a00b3b7cf3d66523d743f7fdddd4c5
GitOrigin-RevId: 245412aa09d917d5f4b6cae13b1b067ba0399601
Change-Id: I9c3a61caff5c717d9005d04c298782a3e330d691