[roll] Roll infra/recipes [sdk] Add placeholder property for enabling Bazel SDK
The Bazel SDK will be used only in conjunction with the core IDK, but
not other niche flavors at this time. This flag is a placeholder that
will eventually skip the Bazel SDK path when we do not expect it to be
built by the subbuild(s).

Also add various TODOs for other cleanups in the proto while we're here.

Original-Bug: 126262
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/868037
Original-Revision: ebc54c4bf593c8e83436be5456219cc0eabc919f
GitOrigin-RevId: df4168d6f8a18ca51fb654639da94258b1b455cb
Change-Id: I3abcf55ad3d220afd875be16a6de5d6cc2ece9f9