[roll] Roll fuchsia [archivist] Support sending logs to serial
This change adds support to write to serial in the Archivist.
At the moment, the change is a no-op, since the structured config is
empty. A follow-up will fill the structured config accordingly and
also remove the equivalent now-deprecated structured config for
console.

This change has a small problem I'll fix in a follow-up.
Given that the Archivist is multithreaded, writing to serial when
ingesting a message can lead to interleaving
zx_debug_write calls when a message has to be split across multiple
lines. Resulting in:

```
00009.317 19891:00000 [netstack] INFO: I'm message one, I'm long and I
00009.319 01103:01171 [mdns] INFO: Starting mDNS
have continuation that ended up in a new line
```

To keep this change a bit smaller, I'm tackling that in a follow-up.

Original-Bug: 100486

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/867196
Original-Revision: 1fb6ecfe780c5c23930f10c09f5fa56e58ad9e49
GitOrigin-RevId: b91980c12a04143aa5bde06760d41c374593084b
Change-Id: I2beccb07add1cb2ce8343e661337f73e9132d542