[roll] Roll fuchsia [botanist] Copy ffx logs to outputs if copying all outputs fails.
Sometimes the ffx daemon fails to shutdown properly leaving a
daemon.sock in the outputs dir which causes an error when trying to copy
the dir to the CAS outputs because it's an unknown filetype. For
debugging purposes, try to copy the ffx logs dir anyway even if we get a
failure trying to copy over the entire directory.

Original-Bug: 128608
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/869276
Original-Revision: 840830a867c67fccbdb5a9471cd256a8ffd505bc
GitOrigin-RevId: 114d4d14db63b8e5cb42d926b33ed373475bb888
Change-Id: Id7f2df70126d6432c9a08c0a8c5fcacb3bc74535