[roll] Roll fuchsia [netemul] Add support for tracing
Add support for running both Netstack2 and Netstack3 with tracing in
netemul and routing the fuchsia.tracing.provider/Registry capability
into the netemul collection so that trace events can be recorded.
Refactor how the debug Netstack2 variant is represented in
`NetstackVersion` to avoid combinatorial explosion.

This added functionality will be used to enable a with-tracing version
of the tun socket benchmarks.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862227
Original-Revision: 34f62e82ddbeaa2f3799c871bfb81cc735921aab
GitOrigin-RevId: b6ea4a3de0e33941e5c780c029cbfb19398bce53
Change-Id: I9b38fdae7d2aaac0caa8929a0820000258ab0422