[roll] Roll fuchsia [starnix] Add SyscallArg
Instead of using a raw u64 to represent syscall arguments, this CL
introduces a SyscallArg type for that purpose. In a future CL, we will
use this type for the third parameter of sys_ioctl, whose type is not
known until we decide which ioctl operation we need to perform.

Original-Bug: 128697
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/869218
Original-Revision: 794e8c4030b2037455aa18a557db935e5e04f6e7
GitOrigin-RevId: 567448406af4fb41c581414275511245aeebe333
Change-Id: I18386fa300950d7c790b98a2cf216965bb1ec8d7