[roll] Roll fuchsia [netlink] Indicate group when sending multicasts
If requested by the application, the kernel is expected to set the
source address of the message. All messages initiated by the netlink
subsystem have PID 0 to match Linux. The group field is set to the
multicast group the message was directed to (if the message was a
multicast).

Original-Bug: b/283827094
Cq-Include-Trybots: luci.fuchsia.try:fuchsia-coverage
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868681
Original-Revision: a89471755a5d0d25b356e15628474538782b882a
GitOrigin-RevId: 37d20be1bbd01560d7484a1dce059ab854a7e0bb
Change-Id: I9fb0430f957391284dc1f35679d50fa10497d57d