[roll] Roll fuchsia [dhcp/client] Implement Renewing state
Implements unicasting renewal requests to the DHCP server we originally
got the lease from and handling the response. If we never hear back from
the server, then we transition to Rebinding, but the Rebinding state
itself is still not finished, so instead we transition to Init as if
the lease had expired.

This means that we have worse behavior when there are multiple potential
DHCP servers on the network, but still allows us to renew "correctly"
in most cases without ever needing to hang onto a lease beyond its
expiration time.

Original-Bug: 81593
Original-Fixed: 124593
Original-Fixed: 125443
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851800
Original-Revision: 2656601f9fa5ff362415ca5c9a4c511f4a92cf5a
GitOrigin-RevId: 65b6362c83f1f307cf921534e40510f0abce32f5
Change-Id: I00a2870a907476bdc9a0efbc0cab0443ed5422dd