[roll] Roll fuchsia [net][e1000] Compile fuchsia support as C++
This is a step toward removing the legacy C bindings for the MMIO
library. The conversion to C++ is very soft touch to avoid disrupting
working logic as much as possible.

Changes:
- Renamed to .cc
- Fix adapter casts from ctx
- Fix device_ops order
- Convert mtx_t to std::mutex
- Fix various void* casts that were implicit
- Remove goto from eth_tx()
- Handle void* arithmetic via a PtrAdd() helper
- Use fit::defer to replace goto logic in bind
- Add CDECLS as necessary

There are no tests, so verification is done locally via `fx test` with a
QEMU instance and e1000, and CQ.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/869516
Original-Revision: 662b3395a6876a7dae439ecf6d7ca44df4b067cb
GitOrigin-RevId: 7f7aa8238781195fa5515ac9c184a9fe752bab07
Change-Id: I584f24e3ef9650a6815276de23aae4e901d308a7