[roll] Roll fuchsia [audio_core] Wrap perftest::ResultsSet in unique_ptr
The smart display hwasan builder is detecting this as a leak. Since this
is used throughout the lifetime of the program and the mem is technically
reclaimed once we exit main, let's just wrap it in a smart pointer to
keep lsan happy.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/869485
Original-Revision: b61aef5895551b4e696e2df6009ac1b34d00ccf3
GitOrigin-RevId: d1f8005bd0beb78163603cd41386fa74fb9d7765
Change-Id: I503c3b05b62ee87edfea1f40337c493c3fab555c