[roll] Roll fuchsia [kernel][riscv] unmask specific external interrupts earlier per cpu
Unmask software (IPI) and external interrupts earlier in the boot
process per cpu to make sure by the time they're needed the cpu is ready
to handle them. This is safe early on because the sstatus.sie bit acts
as the main overriding mask, which isn't set until the cpu is ready to
schedule.

This fixes a wedged kernel boot that happens occasionally where the boot
cpu gets into a place where it is blocked waiting for timers or IPIs to
kick the scheduler, but it never occurs.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868886
Original-Revision: 6b904aa98996a91f1bc7117faa070463feb62239
GitOrigin-RevId: 610ce64cee6678eae5caeb0bee3072726040c0a1
Change-Id: I46301d59fa3c84c0258ad6380b2bd57132d87a3b