[roll] Roll fuchsia [bt][obex] Define type for the `Type` header
The Type Header is used to describe the type of Obex object. It is
based on a fixed set of media types (MIME) and is represented as a
null-terminated ascii string in Obex. This differs slightly from the
canonical OBEX string which is represented as a utf-16 null-terminated
string.

Define the MimeType for representing this header and some conversions
to/from a typical Rust string.

Original-Bug: 120012
Test: fx test bt-obex-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/867458
Original-Revision: b68f9384cee3d813c53f9cb2fc59cb4f4dfb988f
GitOrigin-RevId: f1326c8d3924b164fb0304652581bce88ca20871
Change-Id: I5cb5eb33c9c7bc6d9771fecef207c4651c2c0ea8