[roll] Roll fuchsia [fidl] Require explicit unknown interaction keywords
This changes the fuchsia.git build to require all protocols to be marked
open, closed or ajar and requires all methods to be marked strict or
flexible. This is just in-tree and won't affect the behavior of fidlc in
the SDK.

To allow this change all in-tree FIDL files have been updated to have
those required keywords. The current default or closed / strict has been
made explicit so that the API and ABI won't change.

This will allow us to change the defaults later, keeping the ABI and API
unchanged.

Original-Bug: 88366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868166
Original-Revision: 114c8544fc1e9601b8245218686e172a78e43c45
GitOrigin-RevId: d04e076221a1d4b87ced2646ce78f7ea35df9547
Change-Id: I40247b02b44fbd718e7952769a9da898ac6cb900