[roll] Roll fuchsia [dart][inspect] Remove test
As part of the work to move dart OOT I think we can remove this test.
This test depends on an in-tree Rust component. Rather than rewriting
that component OOT we can remove the test.

The library that the test is testing is only used by the inspect dart
integration test. That integration test arguably serves as a test for
the library itself given its current limited functionality and use.

Original-Bug: b/240154053

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/870443
Original-Revision: 32a609e694e752489534a45056ed4aa0e475457a
GitOrigin-RevId: 986a99d8a4d4975d1b07f90f40d298c7dbe0b776
Change-Id: I428468b407c8a64063f2c8fcac297e51a1c2944e