[roll] Roll fuchsia [netstack3] Don't apply window scaling in SYN-RCVD
We don't advertise a window size that is scaled, however, we mistakenly
used a scaled window size to calculate the acceptable window for any
incoming segment. This made a lot of segments that should be outside the
window considered by the state machine as inside the window.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/870838
Original-Revision: e65af1a4d128141c644ad54f6e74559481a370a4
GitOrigin-RevId: ed83af2ccfd0f9a24fcdd15141fe3f2c3f02454c
Change-Id: Iaefa126f85b1a8ac74ae085944496f84c5aaad0e