[roll] Roll fuchsia [platform_config][diagnostics] Write logs to serial from Archivist not Console
This change will cause the Archivist to log to serial instead of
console.

The new configuration is based on monikers. The configuration reflects
this based on analysis of what tags map to which components.

Based on previous discussion, we are also extending the set of
components that write to serial by including everything under the
/bootstrap realm.

NO_IFTTT=we don't need these changes in archivist default config in gni which are inherited by archivists for testing and the embedded archivist

Original-Bug: 100486
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/867197
Original-Revision: b77eba776202482762e636b4427cfaa4e1077563
GitOrigin-RevId: 9c2171eeab801f1295b4a11dad73b74717679f57
Change-Id: I242a91e787d4569b3d96f5a30a7e815f6f066366