[roll] Roll fuchsia [driver][logging] Register and use a global logger
The logging API makes it very arduous to migrate a DFv1 driver to DFv2.
Additionally, it's working around a limitation which does not affect the
vast majority of drivers - being loaded into a process multiple times.
This limitation should be remedied by EOY once we have dynamic linker
namespaces, at which point we would want to rip out all of this support
in favor of a global logger anyway.

This is still an improvement over the syslog APIs because the logger
symbol is local to the driver shared lib, rather than in it's own shared
lib. This means it will not be shared with other drivers in the same
process.

Some drivers needed to be updated to remove a now unused class member.
The compat driver was switched to always specify the logger instance as
it is one of the few drivers which is loaded into driver hosts multiple
times.

Tested: fx test compat-lib-test
Tested: booted core.x64-dfv2 image and double checked log attribution
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/859786
Original-Revision: 7cf6769e0971693fc1307c6059826596452c75b2
GitOrigin-RevId: adc2093c7c92f7dacc8150a8e965e0610678711f
Change-Id: I8e858fcefa16e2ff0cb9d7776633b13eac4768af