[roll] Roll fuchsia [network/tests] Tun socket benchmarks with tracing
Add a variant of the tun socket benchmarks that collects a trace during
the benchmarks for further analysis.

Add trace events to the benchmarks to show the duration of each
test group, test case, and read/write operation.

The dart wrappers of benchmarks for the purpose of generating a trace
have been collected into a new file to enable reuse of the existing
function for running the benchmark and downloading the trace.

Original-Bug: 125658
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862228
Original-Revision: 23d37f3d43bc57350453ea7c60fb50908260d92f
GitOrigin-RevId: 3629704d036130bf3b7aacb6e7d55d52cb7cfaf4
Change-Id: I51e7ae0c8723d164c8fee43fd45d34400970df70