[roll] Roll infra/recipes [rust] Pass correct host sysroot to compiletest
We had the right idea previously, we were just passing the target
sysroot instead of the host sysroot. This caused proc macros to depend
on libc.so instead of libc.so.6, and when loaded into rustc to compile
tests, it triggered an "Invalid ELF header" error because libc.so is a
linker script.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/871260
Original-Revision: a2b6377a0edae42f119643907146c01921757959
GitOrigin-RevId: ecf65bf1ff5289102a66d5d822b613dc0ecfa3c3
Change-Id: Ia21cfb6f4334a633acdc8448c598ee831db116ca