[roll] Roll fuchsia [debugger] Catch failures stepping through PLT.
When there is a problem with the breakpoint that is step throught a PLT
call, currently we'll just resume execution.

This patch adds a breakpoint at the return address of the PLT call to
catch the various errors that can occur handling the PLT trampoline.
While this ends up stepping over the PLT call rather than into it, at
least execution is not implicitly resumed. In most places we step over
PLT trampolines and something fails (see bug 128935) the user is
actually trying to step over anyway so the problem is not visible.

Original-Bug: 127391

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/869899
Original-Revision: cf8f3a2a5540429496ed23782d8f3d7d4c65e5b9
GitOrigin-RevId: 88b4a26b72cbeabea8147921175f3be5c3558179
Change-Id: I0aa31b91795e05c6154814ad9cc4bccf780989dc