[roll] Roll fuchsia [starnix] Send SIGALARM in setitimer() and implement alarm()
This updates out itimer implementation to generate SIGALARM signals when
firing and also to properly compute the remaining time value when the
timer is queried. This is implemented by running asynchronous tasks on
the kernel thread pool's async executor and using the async executor's
timer heap to track the different timeouts.

Only ITIMER_REAL is implemented as we do not have a mechanism for
tracking the clocks needed by ITIMER_VIRTUAL or ITIMER_PROF. Querying
either of these timers succeeds and returns the disarmed state. We also
support 'setting' the timer to the disarmed state as this is used by
gVisor tests during setup to suppress SIGPROF signals.

This also supports the legacy x86_64 alarm() syscall which is similar to
setitimer(ITIMER_REAL, ...) except that it has slightly different output
behavior.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868194
Original-Revision: 20689ba8369a944700703d6dfa6b449d5f7517d0
GitOrigin-RevId: 5ca6ea98a4a019705d3eec96c2e046c48243f171
Change-Id: I6661911cbb2f539d363eb10a1c5ff71b1da2406b