[roll] Roll fuchsia [netstack3] Fix a TCP data corruption bug
The readable region of the send buffer is only a portion of the entire
buffer, previously we bump snd.nxt without considering short reads are
possible on send buffers, this leads to wrong sequence numbers. It is
uncovered by a performance degradation due to retransmissions because
this bug caused gaps in the sequence numbers which are supposed to be
next to each other, it turns out to be a correctness issue.

Original-Fixed: 128850
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872016
Original-Revision: 13ce82c3023b287ed398d4df14737af072b21cd3
GitOrigin-RevId: 7172f1447a9f130efabbcddd2ff8972f1c35f02c
Change-Id: I60ee0c3f9a7290adb95814e7e7cc0e64a52b48e7