[roll] Roll fuchsia [fxblob] Make target image size optional
Allow setting target image size to None, which means "make it as large
as needed".  In practice, use a similar heuristic that was used for FVM,
which is to make the image 2x larger than the base contents.

This is useful for emulators where there isn't practically a limit to
enforce and it is tedious to maintain an artificial limit.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/871996
Original-Revision: 7e8d6cf2aacdccbd9d1a2e0d01a978c9a965cbd8
GitOrigin-RevId: 0e866e45a4dc50766da8338315468fb530fd08e9
Change-Id: I2db1dfcdb9ba623353f7a2812d62f97bc84d2975