[roll] Roll fuchsia [pkg] Delete fuchsia.pkg/PackageCache.Open
Open depends on the dynamic index which is being removed.

Clients are migrated to Get which will succeed as long as the blobs
are cached, regardless of the state of the dynamic index.

As a result, the cache fallback for resolves will now succeed as long
as the blobs are cached, which could be observed in the case where
ephemeral resolution of a cache package succeeds and then, without the
device rebooting or GC being triggered, a later ephemeral resolution
fails. This behavior change was approved in RFC-0217.

Original-Fixed: 124713
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827996
Original-Revision: b680f2c19d6e699419a52e4ff89581ea3ef75bba
GitOrigin-RevId: 39bf0ac16ed3a0106d88b6558b670437ecbaf343
Change-Id: I55654b0ebeaf79e94efa6672e54eea399d6b360a