[roll] Roll fuchsia [storage-benchmarks] Support Fxblob-based systems
When there is no FVM available, the storage benchmarks will attempt to
create an FVM in the system's GPT and use that FVM for benchmarking.
This supports the Fxblob use case (since there is no FVM there).

The test will not fail if both the FVM and GPT are absent, but no test
cases will run.  An error message is printed here explaining how to run
these tests on an emulator (which is useful for local development and is
typically the cause of neither an FVM nor GPT being available).

Original-Fixed: 126938
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868941
Original-Revision: 1adaab1b91619d22f5657b72482f2b7256eb8420
GitOrigin-RevId: 6e34409bfaafbfc642528585b53c58c59864561b
Change-Id: I51b96d8b760f381a53801f89620215734e74bb4f