[roll] Roll fuchsia [gvisor_syscall_tests] Skip ClosedWriteBlockingSocket for ns3
This test has been flaky for a while despite workarounds. The fact the
test is dependent on the concrete implementation of the socket buffers
makes it hard to device a method for all platforms (linux, gvisor, ns2
and ns3). A previous attempt that accommodates fuchsia netstacks caused
flakes in gVisor tests. Note that this test is also skipped for ns2,
and the waiting behavior on fuchsia happens in fdio, we can skip it for
ns3 for now.

Original-Bug: 128641
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872020
Original-Revision: 133cec798ff1a2cbf2113defd69798697ff05981
GitOrigin-RevId: e611f7e3e191b51e7fa60bc947c2197274c27d09
Change-Id: I98804524111a65d5ea0ce98a9f43b39e670e1489