[roll] Roll fuchsia [boot-shim] Template UartItem and UART_ALL_DRIVERS
This change is meant to allow boot-shim-tests to be compiled with
UART_ALL_DRIVERS=1.

Before this change, when the test would be compiled with
UART_ALL_DRIVERS=1 it would crash due to ABI incompatibility between
what the test think uart::all::Driver is and what the library sees.

By templating over uart::all::Drivers, the mangling (even as the alias
uart::all::Driver is the same, the underlying type is different) is
different, so the generated functions are different and the template is
instantiated with the correct types.

Original-Bug: 129047
Original-Bug: 129046
Test: fx test -v -o boot-shim-unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/870897
Original-Revision: 26a6703529ddb306997ef0ec6aec5ed523188ea4
GitOrigin-RevId: 96dd9fdd1a02e6ce8b081d1f284f8f6db98bc963
Change-Id: I8ac8f46a61c008f584563fc544ba05e0cfb6d40f