[roll] Roll fuchsia [minimal] Add buildbot bundle for minimal
Add a bundle for minimal buildbots similar to other products.

Include an "additional_build_targets" group that can be used to build
other reachable GN targets that are not dependencies of the image and
thus are not built by `fx build` without the bundle. The initial set
of targets is a manual selection of targets that are known and build
successfully on `minimal.riscv64`.

Original-Bug: 128868
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/871619
Original-Revision: 38c23148194421c12c9cd2b2b7115ff45f5c28db
GitOrigin-RevId: 5f9d413a255c273a6048db1f621b6dfa5c6f3b8c
Change-Id: I3e8789fb3651c7152238da1273b51a18ee703741