[roll] Roll fuchsia [starnix] Don't hardcode iface name in `SIOCGIFINDEX`
Before this change, `ioctl(_, SIOCGIFINDEX, _)` returned a value of of
`1` if the requested interface name was `sta-iface-name`. The value `1`
is usually assigned to the loopback interface has a different name and
the name `sta-iface-name` may not be present on all systems.

The changes implements `ioctl(_, SIOCGIFINDEX, _)` (for inet sockets)
by requesting the interface index for the interface name from the
netstack.

Original-Bug: 128604, 129059
Original-Bug: b/286855878
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/870847
Original-Revision: 54b6c47b44c4ee79858cd60df4933ee66c50a2e9
GitOrigin-RevId: 3fdeb7c8a96371f77194b9d2ddc2f3c7faff1789
Change-Id: I12d7df359cf58b91687212cf9ffdb586d5ace719