[roll] Roll fuchsia [starnix] Refactor execve arg reading.
This CL adds a method to memory manager to read a C string into a
vector. This is intended to be used for reading strings that can be very
large.

The other alternative, `read_c_string`, required the entire buffer to be
preallocated all at once, and copied all that memory each time.

The `execve` `argv` and `envp` parameters can take up to 32 pages each,
which was not a good fit for `read_c_string`.

There is still some room for improvement to avoid extra copies and
accept uninitialized data, but out of scope for this CL.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/871261
Original-Revision: c19d644792cf1a0a5d177069305181bfdbb2429a
GitOrigin-RevId: 1e869c199ee44ce55d8f1b1348eefa329962749b
Change-Id: I8fbb419c6c8af11354c42b000ee68477cf91b133