[roll] Roll fuchsia [starnix] Add vdso tests
These tests match the vdso behavior in Linux: almost all memory mapping
operations are allowed on the vdso, as long as they don't split the vdso
mapping.

Currently in Starnix the vdso is only one page, so that test will be
skipped. In Linux it passes.

The CL also adds tests that make sure that the vdso is not writable
across process boundaries (vdso modifications are not visible after
fork or after execve).

Original-Bug: 128842
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/870718
Original-Revision: 14d88b15e611849952188daa059fa54b0742094c
GitOrigin-RevId: 4787f1a01d7dce0f5a5f2e786b782127d75b939d
Change-Id: Idb5c5e446a3bc2d38ddfde9e173f4fa14dcf4033