[roll] Roll fuchsia [profiler] Drop dependency on inspector
This cl changes to manually invoke elfsearch instead of going through
the inspector api which awkwardly wrote out to a FILE. By manually
collecting the information we can now directly write to the socket and
don't need the temp file.

This also helps us have the symbolization information in a structured
format for if we change the output to be structured instead of in
symbolizer markup.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862881
Original-Revision: dec5edef8374fdf0658d43fc968dd0a04473bd40
GitOrigin-RevId: d055b03ff12cdd175a17def4b132213955ac3850
Change-Id: I8e0cc33779bcf89f714cfaae51ce9e7c84053b4f