[roll] Roll fuchsia [boot-shim] Bootstrap '/chosen'
For devicetree item, the early stage where no allocations are allowed,
require use to bootstrap memory layout, cmdline and the
uart(diagnostics).

Thist CL, introduces a stub for the non boostrapping phase, and
an implementation for the bootstrapping phase.

In order to unit test this item in all platforms, the
uart::all::KernelDriver was slightly modifed in order to not assume
uart::all::Drivers always. It was left as the default option.

Introduce MaybeCreate for devicetree binding

Devicetree device nodes specify a list of strings that represent devices
which they are compatible with. The property name is 'compatible'.

This work is required to instantiate the uart driver from the devicetree
blob.

Original-Bug: 78675
Test: fx test uart-unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/860576
Original-Revision: e68f08e5eb16108039e953775f94da60138dab3a
GitOrigin-RevId: 7a739d419646a5bd208e83335c0776e6535d2425
Change-Id: I30d5df48de74d566569d906c2e7c0d9cd6f5a75f