[roll] Roll fuchsia [starnix] Refactor async dispatcher in remote_binder
1. Read the fidl command concurrently instead of sequentially.
2. Await all FIDL related command

This ensures that all fidl related tasks are awaited on the local
executor and allows to remove the cleanup method from the state.

This also solves a crash in the executor happening when the fidl task
was sent to the client task thread to execute the binder command and not
awaited.

Fix: 129211
Original-Bug: 129183

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872496
Original-Revision: 32a8e8d1fe64bc2f60ec7403facf48a88321d5ed
GitOrigin-RevId: 5ca40eb4066f9cfc8b8e121a063b7a29dcfb8639
Change-Id: I2811f139ef1600d4644d1076790274e03a285c2c