[roll] Roll fuchsia [starnix] Delay service requests until UTC clock starts
This waits for the UTC clock to start running before Starnix starts
serving requests and instantiating containers. Many Linux waiting and
timing APIs default or only provide operations in terms of the realtime
clock and do not function in a reasonable way when it is stopped. The
UTC clock is typically started by the timekeeper service early in boot,
often before Starnix itself can start, but in some situations the
startup order can be reversed leading to surprising behavior. To meet
the general expectations of the Linux world this just avoids doing
anything at all until the clock is going.

When Starnix is waiting for the UTC clock to start it emits a WARN level
log. Most of the time Starnix will not block at all and not log
anything.

Original-Fixed: 126111
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/871561
Original-Revision: 57710ab0b8d41a1401a278a80e464dbde6e3ae3e
GitOrigin-RevId: 5113cc96561c6a9ba6b52410226b5323c16719ce
Change-Id: I0d038bc014e94be1c208018ae654b398d0c50159