[roll] Roll fuchsia [trace] Manage rust trace contexts via RAII wrapper
Some specialized trace producers may want to be able to avoid the
computation to compute trace arguments, events, etc when tracing is not
enabled. For this purpose, it is more natural to try acquiring a
context and use that context to write the record after said computation,
rather than first checking if tracing is enabled for the category and
then later obtaining the context which would result in a second check
(and may also give a different result from the first check).

Original-Bug: b/271467732

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/869159
Original-Revision: 9721cce5cb8ca87e5ea468e21e7e82970415ae5c
GitOrigin-RevId: d83f6d669f5c2213570324d0d82595cf54210a08
Change-Id: I184a885fd1ef267d93edc556cabbba25f5827de4