[roll] Roll fuchsia [reland][starnix] Send SIGALARM in setitimer() and implement alarm()
This is a reland of commit 20689ba8369a944700703d6dfa6b449d5f7517d0
which was reverted in commit  83f6fbb7ea390c8bf040cb9fd417783d06ea6494.
This reland fixes a bug in the computation of the remaining time out
parameter for clock_nanosleep. The initial commit was computing this
time in the monotonic domain which resulted in a value that was too
small if the UTC clock was running faster than the monotonic clock. This
reland computes the remaining time against the originally specified
clock (UTC or monotonic) and adds a unit test that uses a injected clock
to catch this bug regardless of the system's UTC clock configuration.

This updates out itimer implementation to generate SIGALARM signals when
firing and also to properly compute the remaining time value when the
timer is queried. This is implemented by running asynchronous tasks on
the kernel thread pool's async executor and using the async executor's
timer heap to track the different timeouts.

Only ITIMER_REAL is implemented as we do not have a mechanism for
tracking the clocks needed by ITIMER_VIRTUAL or ITIMER_PROF. Querying
either of these timers succeeds and returns the disarmed state. We also
support 'setting' the timer to the disarmed state as this is used by
gVisor tests during setup to suppress SIGPROF signals.

This also supports the legacy x86_64 alarm() syscall which is similar to
setitimer(ITIMER_REAL, ...) except that it has slightly different output
behavior.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872261
Original-Revision: a65be31a9cf88532f7a872bc88d984031975843a
GitOrigin-RevId: 5b65d3e1bdab6dc82e452d28d2066c0292290df3
Change-Id: I2ae49c9fad0d06559e99320570a2774ceba21009