[roll] Roll fuchsia [boot-shim] Introduce DevicetreeMemoryItem
This device item is in charge of digesting memory items within a
devicetree. That is, memory nodes, reserved-memory nodes and memreserve
ranges.

Test include individual case parsing, and a mixed case that includes
checking the verified ZBI image.

This Item is meant to be used as a bootstrapping item, that will allow
general matchers to perform allocations.

Original-Bug: 78675
Test: fx test boot-shim-unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858118
Original-Revision: 6227d6a487163aa36955081e057d34910214d3d3
GitOrigin-RevId: 0857833bbcff5f36a4040d1b3af6e653a7827e78
Change-Id: Ia4fc0c53d3b0614b6e8d8dc86452b274091d9efe