[roll] Roll fuchsia [build] Fix rust_proto_library() hermeticity
The rust_proto_library() template invokes the `prost-build`
Rust program that does not generate depfiles at all, so this
CL wraps the call with a wrapper script that can compute the
list of transitive protobuf imports.

This allows generating a proper depfile, and verifying that
the `packages` value passed to rust_proto_library() is also
correct. If not, a human-friendly error message is printed
explaining the issue.

Original-Fixed: 127778

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/873217
Original-Revision: 1094f6c74f0978ecc9c7900bd15c8e38883c1dca
GitOrigin-RevId: 8e22a2cd3df483abd29fa5d65240969b0e93410a
Change-Id: I9f12ffef33d604d93bf2089da6e9f7c795934ce7