[roll] Roll fuchsia [starnix] Add a minimal tracefs mount type
This adds a tracefs file system that contains only a single node:
trace_marker. Unlike tracefs on a Linux system, this implementation does
not maintain a ring buffer of trace events, so it does not expose trace
events via readable files.

When data is written to the trace_marker file, if the data is in the
format used by atrace, it will be translated to a Fuchsia trace event
and written to the Fuchsia tracing system.

Original-Bug: b/271467732
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/869977
Original-Revision: 66dc6da8c8b6f80d9cc3c1be994a233d8847f2fb
GitOrigin-RevId: 0db0714d32117f8fdb98d686f106865a5bcde89f
Change-Id: I83b8f6652f2245b766de235aa24b2864b2126b39