[roll] Roll fuchsia [fxblob] Add an as_blob option to Volumes.Create
Volumes.Create should not rely on the name of the volume (i.e. checking
if the name is "blob") to determine how to mount the volume. Strings
are a weak API. Instead, Volumes.Create now takes in MountOptions which
contains an as_blob option. When as_blob is set, the volume will be
mounted as a blob filesystem.

Original-Fixed: 126745
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/870176
Original-Revision: 496dc0af2a55995b0700780346cb8ecac38bade9
GitOrigin-RevId: 15fab47157222373d0e50b75b69c521a697950c2
Change-Id: Idef8ac0b4562a153e4c865b14ad0668b0b3166bd