[roll] Roll fuchsia [check-licenses] Enable license-classifier.
Previously, check-licenses used simple regex pattern matching to
detect licenses in a text file. This works, but setting up and
maintaining the regexp patterns so they work for every project is
painful and error prone.

https://github.com/google/licenseclassifier is a library that
can generate a similarity score between a given blob of text and
a known license text library. In practice it is very accurate,
fast, concurrency safe, and the library of license texts is
significantly easier to generate and maintain.

This CL flips the switch from the old strategy to the new one.
* Remove references of the "license" package across the project.
 -> that package implemented the regexp solution.
* Adjust allowlists to temporarily allow a handful of new issues

Original-Fixed: 113024, 111682, 102068
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868757
Original-Revision: 45032cf537e75e41a796c7981950ca178485690b
GitOrigin-RevId: 5aa6bfa2b7d5b6b4e4a0b8dcfc5e18da1f497137
Change-Id: I7512e0a0e225576ed02c332a48b67fb30775dfa2