[roll] Roll fuchsia [fuchsia-controller] Support isolation.
This adds support for an isolate_dir object to be passed to the fuchsia
controller context. When this isolate dir goes out of scope, the
directory it points to will be removed to clean up the FFX daemon (the
ffx daemon will shut itself down once the socket it's pointing to is
deleted).

Original-Fixed: 129193
Multiply: fuchsia_controller_e2e_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872267
Original-Revision: f079a45259b2ffaffd2c52c2fa6fd507637003a1
GitOrigin-RevId: d0d01acdfcc2a7300dc7ea6f0681d142338d4a49
Change-Id: I37610930fd65d85f792b11a04fb64efe7c76f8df