[roll] Roll fuchsia [fxblob] blob_writer stops splitting writes if remaining bytes fit
The blob_writer client used to split all writes so that none would be
larger than half the VMO to prevent filling the entire VMO with a
single write (in which case the server would need to wait for a FIDL
round trip in between processing writes).

The blob_writer client no longer splits writes if the VMO is large
enough to contain all the remaining unwritten bytes.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872725
Original-Revision: 8bbd261331a020beb729892427902dc47d5ee093
GitOrigin-RevId: b10097f37ae9a6e7b3a087944822d707e0499147
Change-Id: I8464e949ffc52b9a2ae2232da732e923cf7230b3