[roll] Roll fuchsia [starnix] Fix remote binder task creation
This fix remote binder task creation.
Before this CL, if a request was received while a task was considered
unassigned, and an unassigned request was available, it was possible to
take over the unassigned task, but no request a new task for the already
present unassigned request. This lead to deadlock where some unassigned
task where never handled.

This CL fixes the issue by not adding a task to the unassigned pool is
there is an unassigned request. Instead the task is assigned to the
request. This ensures that if a new task is added and an unassigned task
is present, there cannot be any already present unassigned request.

Fix: 129183
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/873237
Original-Revision: 84a0dc66c8c9261c50453eee42db6687fe81d151
GitOrigin-RevId: 059304c9ce83eebccedc25c20d64c450f318c156
Change-Id: I52532aef9b2cf5095ed1699aeea0d02a4849aaad