[roll] Roll scudo [scudo] Change overloaded function name append.
The ScopedString class has two functions named append. One takes
a va_list, but on some platforms va_list is typedef'd to char*.
That means that this call:

  std::string value;
  Str.append("print this string %s", value.c_str());

The compiler can incorrectly think this is the va_list function,
leading to crashes when calling this. To fix this, change the name
of the va_list function to be vappend to avoid this.

Fix https://github.com/llvm/llvm-project/issues/62893

Reviewed By: Chia-hungDuan

Differential Revision: https://reviews.llvm.org/D153389

GitOrigin-RevId: 240c4e1e16c215a7b2704eea68f54aea68d19c7f
Original-Revision: 405d87b7408ebac684a29b90bff2d9dc2c39a2fc
Change-Id: I7a3d212a247b258372c2896491c9407e8378d368