[roll] Roll fuchsia [zbi-format] Introduce v3 CPU topology definitions
This change introduces the v3 CPU topology definitions from FIDL.
A follow-on change will add support for them within the platform.

The v3 is being introduced to be layout-compatible with that FIDL can
represent - but presents information equivalent to the v2 in a similar
form. In particular, we take advantage of the experimental `overlay`
FIDL feature to represent discriminated unions.

Original-Bug: 111453
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868172
Original-Revision: 46fa663a439a5385643c53d3770c052dbca8d7e4
GitOrigin-RevId: fe2f2b68781910d9e49c46187a29edfef0872ba1
Change-Id: Ia3c02af0ff5fd374379c845b0f9da214f29eb150