[roll] Roll fuchsia [storage] Replace BlockDevice::Controller with Rebind
Instead of giving clients access to the controller and asserting
if it doesn't exist, provide a targeted API to rebind the device.

This way, we can remove reliance on FIDL multiplexing and also
return ZX_ERR_NOT_SUPPORTED if we don't have access to the controller.

Original-Bug: 127870
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/874920
Original-Revision: 57f8aa64255b5c34c8c25453010cc343e2329295
GitOrigin-RevId: 474bdb0ca0d4cb66c6bb15f4d5a8306cb899e864
Change-Id: I1e90daedefd3526890b2b03aaa12fb62360aa056