[roll] Roll fuchsia [fidl][server_suite] Migrate to packages and subpackages
This change refactors the server_suite tests to leverage packages and
subpackages.

The goal is to be able to move the Dart tests OOT.

This approach (already taken by the inspect validator) allows us to do
that by putting the harness package in the termianl product. This will
come in follow-ups.

Original-Bug: b/287586341

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872476
Original-Revision: 665092b980127be24e8bf40682ac44391bd85ee0
GitOrigin-RevId: b8fec29f1445ef1a32da90d383e68f03c0acb2c8
Change-Id: Ie37be63c44ac35a59c15dd454680c14b3c92be48