[roll] Roll fuchsia [fidl][client_suite] Migrate to packages and subpackages
This change refactors the client_suite tests to leverage packages and
subpackages.

The goal is to be able to move the Dart tests OOT.

This approach (already taken by the inspect validator) allows us to do
that by putting the harness package in the termianl product. This will
come in follow-ups.

Original-Bug: b/287586402

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/871517
Original-Revision: 4483ca886270ca1f4e8eb8d510cc40c2151167e1
GitOrigin-RevId: 9df86c4dcc055afa8d169ad95f41f5c7f7d3b7f5
Change-Id: I62bad2aa02e55aa9773f6937bb51ae9b910e252c