[roll] Roll fuchsia [starnix][kernel] Move most capabilities to optional besides Vmex
Concretely, what this does is that the parent of the starnix kernel
component could route void to these optional capabilities. This would be
useful if a test doesn't exercise that capability, and would like to
avoid routing warnings.

If starnix did use a capability and it was offered from `void`, that
will result in a DEBUG level log which is hidden in tests most of the
time. The operation that depended on that capability will most likely
fail.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/875463
Original-Revision: 8ff019f21a47b0c9797adb20db1ef675070e7e7c
GitOrigin-RevId: c5c029e92105fee1aafdfad05623cc08cf96d117
Change-Id: Ia35ee5116d8e0659aaa38c7daf362e89b97a3146