[roll] Roll fuchsia [debugger] Cleanup downloading retry logic
This change modifies the way downloads are performed. Most visibly, we
no longer check a server for a file before attempting to download it.
This added complexity with little benefit, since if a server doesn't
have a file it will simply return a not found error immediately, and if
it does have the file we want to immediately download it anyway.

Second, a second callback parameter is added to the Download object,
which allows it to report back when it has run out of servers to try a
download against. This is particularly helpful when downloads are kicked
off immediately at startup (like when debugging a core file) and some
symbol servers have not finished authentication yet, even though the
system knows about them. This way DownloadManager knows how long to wait
before emitting error messages to the console about missing downloads.

DownloadManager is now the sole owner of Download objects. Instead of
sending around std::shared_ptr's to the Download methods, which
complicated the lifetime ownership of a particular download. Using the
new callback mentioned above, DownloadManager is able to determine when
a download is sitting idle and there is no more asynchronous work going
on that could make that download succeed.

To facilitate the correct ownership of each Download object,
DownloadManager now subscribes to all SymbolServer events through the
observer interface, rather than having a special method called from
the system.

Because of the above, DownloadManager no longer needs to keep track of
failed downloads. If a particular download failed, it's because it
exhausted all available symbol servers. If a new server is added from
the command line, and the process is still attached, it will kick off a
new download against the new set of servers (which will try the servers
that already failed). This could be optimized to _only_ kick of
downloads against the new server(s), but we're typically attempting
downloads from < 5 servers so that level of optimization is not
required yet.

Add unit tests for DownloadManager. This required a small change in the
SymbolServer base class, which was previously not having the
authentication flow tested in multiple unit tests. The static path
variable that depends on `$HOME` environment variable set in
symbol_server_unittest.cc would only work if the symbol server unittests
ran before the tests in download_manager_unittest.cc.

Original-Bug: 61746
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868942
Original-Revision: 1d75daa06f690988b067e5cb2ab41cd10294027b
GitOrigin-RevId: 68ad8ff3146576c163b3c78e16181a2737bbc65f
Change-Id: Icc4da518a97e00709b774055a55001282b3548bb