[roll] Roll fuchsia [tracing] Add Additional Help for Connection Errors
Added a small blurb with some possible diagnosing steps and a pointer to
the tracing documentation when tracing fails to connect to trace
manager.

When it fails to connect it now prints out this message:

```
Error: ffx trace was unable to connect to trace_manager on the device.

Note that tracing is available for eng and core products, but not user or userdebug.
To fix general connection issues, you could also try:

$ ffx doctor

For a tutorial on getting started with tracing, visit:
https://fuchsia.dev/fuchsia-src/development/sdk/ffx/record-traces
```

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/875718
Original-Revision: 1a1f15157340b62ae3ea5ce05c7b4b4ce2c575df
GitOrigin-RevId: fc807529d39187636bde1a2bf0b82414557aa56f
Change-Id: I78fabe64cb80242a2c24d74ef4f7970626bfd07c