[roll] Roll fuchsia [netstack] Convert interface auth grant to proof
Add a helper fn to convert from `GrantForInterfaceAuthorization`
to `ProofOfInterfaceAuthorization` to be used as part of authorizing
routes to be added that use a particular interface.

Jeff and I discovered that it is not possible to create an EventPair
with an `INSPECT` right, and that it is not necessary to request
the basic info about an `EventPair`. Therefore, it has been removed
from the FIDL definition and now `ProofOfInterfaceAuthorization`
does not need any rights.

It is not possible to specify `NONE` as the Rights on the token, so a
bug has been created to update it when possible.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/874337
Original-Revision: 8143b5dc1e0e00faf9b4def4440d25700cce349e
GitOrigin-RevId: 463a0a07dc911ea6039315f094949c1c183150b5
Change-Id: I28c220eaa664f49d545b92791baf65df846b65cd