[roll] Roll fuchsia [pbv2] introduce LoadedProductBundle to track where product bundles are loaded from.
This CL adds a new struct which wraps the ProductBundle and allows us to
track where the product bundle was loaded from even if we used the
configuration domains to load it. This will be used for the emulator
code paths to let the developer know what product bundle is being
loaded. The users of the load_product_bundle method have been updated to
call into() which immediately converts the object back to the
ProductBundle enum. In the future these call sites can be updated to use
the LoadedProductBundle.

In the future the ProductBundle will be converted from an enum to
a Trait so that we do not need to introduce this extra layer.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872789
Original-Revision: 3bccec284d20e343caa0e8dd5afc0daddac9ce4c
GitOrigin-RevId: e2c1e5a2830268ccffcd4e386e694495592c941e
Change-Id: I2f52bf0a8f929a0f0c63dac25e06da4b4a7e2767