[roll] Roll fuchsia [net/mdns] fix cache flush implementation in instance requestor
This CL fixes the cache flush implementation in instance requestor,
which previously cleared all A/AAAA resources for a given name from
its cache when it receives an A or AAAA resource with the cache
flush bit set.

This change causes the requestor to flush all *matching* resources
(A's for and A resource, AAAA's for an AAAA resource) less than one
second old, as specified in RFC 6762.

Original-Fixed: 129021

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872186
Original-Revision: 8c37d59a9faca38c2840c98064bf71fc9e6f027c
GitOrigin-RevId: c3845d1d6383716e2293a213855f3f710e5ca02d
Change-Id: I414d75a755e87252eec49a5bdfbee49bb5897078