[roll] Roll fuchsia [bazel][assembly] Add bootserver_pave_zedboot if recovery is zedboot. This will match the images metadata created from GN assembly
After digging more into the ssh failing seen in
http://go/bbid/8777554067385911313, it looks like the error is
introduced one step before when we pave the device. By comparing the
metadata for zircon-r and vbmeta-r, the difference is the metadata data
created by bazel assembly lack the field `bootserver_pave_zedboot`.

Note we are not dedup the metadata data in this CL. Instead of modifying
the metadata, we are going to avoid creation of recovery image through
GN completely, which will be Jay's fix in another CL.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/876257
Original-Revision: 0dc1dc6c4cb3d313fb386a818b3487398772781c
GitOrigin-RevId: 7b70ec8e2b5e16dc78c43f6806e4f911d0a94922
Change-Id: I33768a5a27a64d6343a51f922771fccdab13b2ef