[roll] Roll fuchsia [wlantap] Ensure child classes are destroyed in order
WlanPhyImplDevice now ensures that WlantapMac is destroyed before
WlantapPhy is destroyed because WlantapMac keeps a pointer to
WlantapPhy, which may result in a use-after-free if they are not
destroyed in the correct order.

This also uses a ServerBindingGroup in WlantapMac to ensure that FIDL
connections are closed when the class is destroyed.

Original-Bug: 129563
Multiply: hw-sim
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/875836
Original-Revision: 5e2bcc462246d65690a098976450cf83f6a98240
GitOrigin-RevId: 03cda3882933226621fc461fbfcb3d3a467d4fa1
Change-Id: I5729d44f7c5d75aec6f3da247f63e4dc39d70b42