[roll] Roll fuchsia [net/mdns] honor cache flush flag for subscribed host names
This CL fixes an issue in which the 'host name requestor' is not
honoring the cache flush bit in address records (A/AAAA). This
bit should cause previously-seen addresses for the target in
question to be forgotten.

This CL also addresses problems with the way addresses were handled
previously. Because cache flush behavior was not understood when
this agent was originally written, there was an attempt to use message
boundaries to determine when to purge old addresses. That logic has
been replaced with proper cache flush handling.

Test: new unit test added

Original-Bug: 128612

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/874541
Original-Revision: 3ce0a37f94969e1e3270a4e7c31384ab699d1c9d
GitOrigin-RevId: 5836cd52c1968d464fc7a66363c97094e16760a1
Change-Id: Idaa4a1d42cf952a29f4616322910c3959b9cf2dd