[roll] Roll fuchsia [driver][sdk][testing] runtime rework
This change introduces a change into how dispatchers are created and run
in driver unit tests. It creates a new DriverRuntime class that is in
charge of creating the foreground and background dispatchers, as well as
running them. It provides many of the same APIs that async::Loop does
like Run and Quit.

Original-Bug: 127188

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/871508
Original-Revision: 5bd4bfa09168ce5aee9bd29ce0ea01899b27eefc
GitOrigin-RevId: d7bcdd68b72d59921bada3608b3150bea950de68
Change-Id: Ibfdf3494990bd0045460235b77783beaf39d70d5