[roll] Roll fuchsia [netlink] Cache interface `Control` handle
...instead of getting a new handle for each interface update request
(like adding/removing addresses).

This change holds the `fnet_interfaces_ext::admin::Control` handle with
the rest of the interface's properties when the control handle is first
obtained. This avoid handle thrash.

A side-effect of this change is that requests modifying interface state
will now fail if the interfaces eventloop is unaware of the interface
(has not received a `fuchsia.net.interfaces/Event` for the interface).
This is okay because if the interface was actually installed, the
eventloop will eventually learn about the device and until then, it will
act like the device has not yet been installed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/875956
Original-Revision: 171daf0e6dc71979b1dfacc514a16a1f250b0fb9
GitOrigin-RevId: 5cd681583babf70e07e76fa385f107ac90ae7afb
Change-Id: Id961e395c8e647646acee0f76f8c706e8114867f