[roll] Roll fuchsia [gvisor_syscall_tests] Skip variants of TCPListenCloseDuringConnect
The two variants are flaky in the absence of dual-stack sockets. The
following can happen:
Assume that the listener binds to a v6 address with port p. Then the
connector connects to v4 address with port p. Now this normally triggers
a RST since no one is listening on port p on the v4 address, rendering
the test to pass (though not intended). However, chances are that port p
is used by some other listener in other tests running, then our test
won't pass as the test is expecting a RST. Besides, the fact that the
test iterates 100 times makes it easier to trigger the flaky case.

Original-Bug: 129660
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/877583
Original-Revision: 74cedc48ddd6cd5e38ea8fd70a2ca1d5bf21e084
GitOrigin-RevId: bd952be2770bbf0d0302b6077a2053cc1cdb0423
Change-Id: I7d064be1f0bbfa576b90242a61ba780a655f70ad