[roll] Roll fuchsia [netstack] Specify TRANSFER rights on `Proof`
In another CL, we found out that the TRANSFER right is required to send
the EventPair across a FIDL channel at all, even if it is not explicitly
being duplicated/replaced.

While not necessary for functionality, the FIDL was updated to reflect
the lowest level of rights necessary to follow the guidelines for
specifying handle rights: see RFC-0028: Handle rights for more info.

Fix: 129584
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/877516
Original-Revision: a76e949c0f42761f8d71c1cc439dbdcaca397006
GitOrigin-RevId: a77cb079a93248cff26403afa4ecee6ba249f62d
Change-Id: Id7281cb41479cde309e4b8189055788c62c21e7a