[roll] Roll fuchsia [archivist] Remove Trie and UniqueKey in repositories
This refactor simplifies the follow-up which gets rid of
ComponentIdentifier (in favor of ExtendedMoniker) and simplifies a
lot of code.

As part of this change we can now delete the "UniqueKey" which was
needed only for CFv1 purposes.

The Trie was introduced at the time for two reasons:

1) Components in the v1 topology could share the same moniker and we'd
   distinguish them by the instance id. We no longer have v1 so we can
   stop supporting this.

2) Do not repeat the key multiple times. Since then, we now have string
   interning via FlyStr so we no longer need that. We can just use the
   ComponentIdentity as the key.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/877437
Original-Revision: df02967884e28a3ff5f10fbd9aded46323159b55
GitOrigin-RevId: 8710c4dce39f16ed1b7713103d6ae7f1b0558739
Change-Id: I47996833a263961618ac0ea0bdb1f57a2276efa4