[roll] Roll fuchsia [paver] Add BlockDeviceClient
Add BlockDeviceClient which represents a block device.
Having `block_fd` on PartitionClient was not helpful, as
most PartitionClients were returning invalid file descriptors.

Getting a BlockDeviceClient from a PartitionClient can return an
error if it's not the right type of client. This is better than
the asserts and casting that was happening before this CL.

Original-Bug: 112484
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/877776
Original-Revision: 3656ce29b3d78b0d655fd3689b97fac152f0d2ab
GitOrigin-RevId: e22a7f89995f08b5308b13e24dee331c1ce8f022
Change-Id: Iacb203a6591f87f59c00783674ff0e3b59ab2b4e