[roll] Roll fuchsia [audio] Fix audio_driver_tests use-after-free
Fix a possible usage after free when a device is detected after
an initial device detection is completed. Related to this, all
device-related timeouts are eliminated, so we rely on the test
execution framework to timeout and kill the test suite if the
device interactions time out.

The DetectDevices method was incorrectly using stack variables
in a callback that could be invoked after the method completed.
DetectDevices is not designed to keep discovering devices after
it completes. However, device_watchers_ are stored to keep the
channel used to connect to the devices, which can lead to
DeviceWatcher callbacks after DetectDevices completes.

In this CL we make sure any subsequent callbacks are treated
as errors. We further safeguard these variables against misuse
in the future by moving them to static storage.

A bug is fixed, related to the number of virtual_audio
instances that are added. This actually was the underlying
issue that uncovered the use-after-free.

Original-Fixed: 129582

Test: audio_driver_tests in asan build
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/877677
Original-Revision: e558dea7476c10ce2ad6f578e3432292511efa4d
GitOrigin-RevId: 86293c3d64f6fa9925de9d229d2a7b3ffd539440
Change-Id: Idc953010489f1e3ed41784035e6281e2b1adb0a3