[roll] Roll fuchsia [fxfs] Build a sparse image in assembly
Generates an image `fxfs.sparse.blk` which is in the Android Sparse
format.  This image is provided in the flash manifest, which is already
handled by the fastboot protocol in a more efficient manner.

The sparse image is significantly smaller than the raw image, since all
padding can be represented in a few bytes.

This change includes substantial refactoring of the `sparse` library, as
well as adding the `unsparse` method (which is currently only used for
testing).

The equivalence of the sparse image to the raw image was manually
verified with the `simg2img` tool.  The expanded fxfs.sparse.blk is
identical to fxfs.blk.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/873758
Original-Revision: 126b57188094802fbc0e58ef15cf6fadfd218648
GitOrigin-RevId: c7c5029e6108dbbd158a390d05b7053bd37e151c
Change-Id: Ic9035e589f45776c254cb7ceca1f62c16bc1188e