[roll] Roll fuchsia [fshost] Remove time-out from find_data_partition
We shouldn't need the time-out for finding the FVM data partition. By
using VolumeManager's GetInfo method, we can be sure that all the
partitions will be present in the topological tree.

This now means that if no data partition is present, we don't have to
wait for the timeout to expire to find out.

Original-Fixed: 129098
Original-Fixed: 129699
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/877236
Original-Revision: 591781a73a805bb1544524e45597c35e2791b6ec
GitOrigin-RevId: 6c152cc028479ea089307206d79a4138b32ea338
Change-Id: I40cf1bb2587eaf79164e0aaa3d17242d03251a39