[roll] Roll fuchsia [sysmem] Fix buffer collection async trace events.
Fix the labels of the async end events to match the begin events so that
the async events terminate and accurately reflect the allocation times.

Also add a trace duration for `OnBuffersAllocated` to make it easier to
locate and measure the activity related to allocation completion.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/875505
Original-Revision: 05409efa7360016c20f7daa8a7c04f07fa2f40f4
GitOrigin-RevId: ec3b0ecf4f0f88d563a0564cf47c1619b46dec06
Change-Id: I0bbeb4ab6af743ee76b970530a29e3d749b2fb6d