[roll] Roll fuchsia [memory_sampler] (Poisson) sampling in bytes between samples.
With the current instrumentation strategy, reporting every single
allocation and deallocation is prohibitively expensive on
moderately-to-large sized processes: the FIDL channel used to
communicate data with the profiler service fills up too fast in the
absence of backpressure, causing the connection to eventually be
severed. To mitigate this, this instrumentation randomly samples the
allocations to be reported, hence providing a way to reduce the
instrumentation overhead at the cost of accuracy.

We use a sampling strategy similar to that of Chromium's and
tcmalloc's sampling heap profilers: Poisson sampling with a constant
average sampling interval, in number of bytes allocated between
samples.

This initial support embeds a constant interval; we intend to
introduce configurability in a subsequent commit.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/864019
Original-Revision: d304cc38419b71cf039791f4c99798173a4bd9b0
GitOrigin-RevId: df8e5c94c99abede9cf85b0485095c1040d8923d
Change-Id: Id17241c21e3fba9c818eccd91ee0730fc7aff17e