[roll] Roll fuchsia [starnix] Refactor task data needed to compute what process can be waited
Before this CL, the ThreadGroup contained a Map of Weak<Task> and was
considering that these weak pointers could always be upgraded because
they were cleared when the Task was dropped. This is incorrect, because
a Weak pointer will become not upgradable before the drop implementation
is run. To fix this race, this CL does the following:
1. Stop considering the weak pointer is always upgradable
2. For computing whether a process is a target for a given wait, share
   some data with the tasks, so that this data is available even when
   the task destruction is in flight.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/878491
Original-Revision: 3f9f596a9374fb5a84c18f987e4a8a7302bde9f0
GitOrigin-RevId: 3624c70916e4e3ed47f4a2f8e1fb5c5f8eed941c
Change-Id: Ic947a61f4dad2047349aa2e04bfa6f99f969b927