[roll] Roll fuchsia [bedrock] Delete many unused types and libraries
We have built a lot of capability types and execution traits as an
learning exercise. However, they're not used in the short term when
refactoring ELF runner to use bedrock, so keeping them alive and working
is just a drag force.

`multishot::{Sender,Receiver}` will be dealt with in the next CL because
it's used extensively in serve_processargs tests. But we'll be replacing
it with `Open` at least in the short term.

We can always resurrect them if we find that we do need some later.

Original-Bug: 122024

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/878915
Original-Revision: 6d788007161ffca4c74e5f429c0ba1d6508822de
GitOrigin-RevId: daa13ab75786afed56e34e66f8110d89a4d58b3e
Change-Id: I45fa6696e29aff0f4650211bbe15075c273fb2a4