[roll] Roll fuchsia [sdk] API Review for driver mirror libraries
API Review to put the following libraries into the partner sdk:
 - driver_incoming_cpp
 - driver_outgoing_cpp
 - driver_logging_cpp

These libraries are the driver counterpart to existing component
libraries. They are mirrored to be as similar as possible to the
counterparts.

Due to current limitations in how drivers are run, these libraries
are not a part of the component versiones of the libraries.

For incoming/outgoing we provide driver transport variants of
everything and regular components do not need to bring in the
driver transport fidl dependency. They also do not use globals like
their component counterparts as drivers can be in the same process.

For logging, the same problem of globals not being possible
currently apply.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/861332
Size-Review: Aidan Wolter <awolter@google.com>
Original-Revision: 09ff9df5f416a17dcee1de5d0a85b7d392290e80
GitOrigin-RevId: 1c133114b5af47e08a37d9545748ebfd3bb48801
Change-Id: I5ded57689b5368303123e56b5aa6aceb5f355d9c