[roll] Roll fuchsia [packet] Simplify execution of Serializers
Previously, `Serializer::serialize` consumed a `NestedPacketBuilder`
which represented the layers outside of `self`. The innermost call to
`serialize` was responsible not only for allocating a buffer which
satisfies the various constraints (especially prefix and suffix length),
but was also responsible for performing serialization itself by
delegating to the `NestedPacketBuilder` provided to it as an argument.
As a result of this design, serialization was more complicated in terms
of code and in terms of call paths than necessary.

In this commit, we simplify `serialize` in the following ways:
- We remove `NestedPacketBuilder`
- `serialize` takes a `PacketConstraints` which represents the outer
  layers, but does not take any `PacketBuilder` or `NestedPacketBuilder`
- The innermost call to `serialize` is still responsible for allocating
  a buffer which satisfies the relevant constraints, but is no longer
  responsible for serializing outer layers
- Instead, on the return path, `serialize` is responsible only for
  serializing the `PacketBuilder` owned by `self` - all encapsulated
  layers have already been serialized by recursive calls to `serialize`,
  and all outer layers will be serialized in the future by callers

In practice, this change results in a significant reduction in LoC in
the `serialize` module.

Note that the API consumed by most users - namely methods on
`Serializer` with an `_outer` suffix - is unaffected by this change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/861377
Original-Revision: 65d001f3cb6f7c7a6fa8ca6f556e4c2205783d29
GitOrigin-RevId: 3ae15eea795582eec62d1dc2b9c905e5848d74f1
Change-Id: I279e223157a3a41904b04013465c0a4a0fb2e711