[roll] Roll fuchsia [packet] Rename and simplify `SerializeBuffer`
Previously, `SerializeBuffer` was the sole argument to
`PacketBuilder::serialize`, and held the header, body, and footer. The
header and footer represent the memory regions into which `serialize`
serializes its header and footer, while `body` is initialized by the
caller to contain the logical packet body to be encapsulated.

`SerializeBuffer`'s name implied that it was a buffer in the family of
the other `XxxBuffer` traits, which isn't true - it was more accurately
thought of as a set of views into another buffer.

In this commit, we simplify this setup by renaming `SerializeBuffer` to
`SerializeTarget`, and by making it so that `SerializeTarget` only
stores references to the memory regions for the header and footer. The
body is now passed as a separate argument to `serialize`. We also
simplify `SerializeTarget` by making its fields `pub` and by removing
some accessor methods.

Finally, having freed up the `SerializeBuffer` name, we rename
`TargetBuffer` to `SerializeBuffer`, to reflect that it is the `Buffer`
trait that defines the `serialize` method.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862010
Original-Revision: b7dd3a6a6d72d0038979b43bcaf906d42817a33c
GitOrigin-RevId: f5984d2c0a16606a875975fffbf65d4638035515
Change-Id: I94f10c7e1aa67ab63ca02832225b8fb3bb61bb4c