[roll] Roll fuchsia [assembly] Only rebase ui.sensor_config path if not already rebased
This is a temporary check for a transition.

If the ui.sensor_config does not start with '../../', then
rebase the path.  If it does start with that string, then
assume that it was already rebased.  This is so that the
usages of this (in vendor repos) can switch to passing the
file in as an already-rebased path.

Original-Bug: b/289396131
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/879711
Original-Revision: dcc15d7495e8982eebfbbcf50d6935a823fc0974
GitOrigin-RevId: 764397c82ebe6be0fb1bc090ef1f0c5754edee2f
Change-Id: I17d488724d1417fc62673694a5b2414a543caef0