[roll] Roll fuchsia [network/tests] Disable NS2 tests in generated src
Include the `#[cfg(not(target_arch = "riscv64"))]` attribute in the
`#[netstack_test]` macro output, not in the source code for the macro.
Since the macro is compiled on the host machine, it isn't aware of what
architecture its generated output is going to be compiled for. The
attribute needs to be present in the macro's generated output so that
when that output is compiled for the target system, it can disable some
test functions for RISC-V.

To demonstrate the fix and prevent backsliding, add some network tests
to the test group run for the minimal product, which will be run on
RISC-V.

Original-Bug: 122609
Original-Fixed: 129748
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/878477
Original-Revision: 6c79a86758f44c2e66f024fc1171494cd6926be7
GitOrigin-RevId: 017e8e76751f8b37eaf22214ac4f28407cc20122
Change-Id: I153e468abeea0f42fea9ad96083eedf74d795c99