[roll] Roll fuchsia [boot-shim] Introduce RiscvDevicetreePlicItem
This item is in charge of parsing the plic node from the devicetree.
For now it only parses the base address and the number of IRQs, which is
enough for what the current zbi plic driver needs.

As part of this change, PropertyDecoder has been extended with
|TranslateAddress| method which recursively translates addresses
until is not able to. This is necessary for all addresses, and its been
observed in devices dts. Since in order for the driver to properly
understand the mmio address it needs the view of the root node of the
memory address. So it needs to be translated all the way to the root if
possible.

Original-Bug: 129970
Test: fx test -v -o boot-shim-unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/879511
Original-Revision: fb3ce5cadb548812046aba3d44a5c05a9052d3db
GitOrigin-RevId: d5dd02fb07e449fa405b1509b99f2b475b7c4cf5
Change-Id: I632f62b5f15c85d43aac3f62515b44f1d84b711c