[roll] Roll fuchsia [starnix] Implement `ioctl(_, SIOCGIFHWADDR, _)`
This `ioctl` returns the link-address for an interface.

The implementation simply translates the requests to the NETLINK_ROUTE
implementation with a temporary socket so that there is only a single
entity in starnix interacting with the Fuchsia netstack.

Original-Fixed: 129987
Original-Bug: b/286855878
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/879691
Original-Revision: bf32a4e861f2bccc78c3745b6f6937abb3ce0b1c
GitOrigin-RevId: 454042372fb8a8bf787a1abe6403192f7e51b53e
Change-Id: Ie7b43a4e146e0f1b8c804de5014294246bf95282