[roll] Roll llvm-project/libc [libc] Remove flaky static assert from RPC interface
Summary:
This function is intended to only be used on the GPU as a shorthand. The
static assert should only fire if it's called ,but it seems that its
precence can sometimes cause issues and other times not. Simply remove
it as it's causing build problems.
GitOrigin-RevId: a6b070c1a16d68f07cb622f902a491d12ae98ae4
Original-Revision: 4d080533bfc44069e849aa8a2188b25dbb95725a
Change-Id: I6a2c405a33939d15e79b53bde4fd0e076c807b2a