[roll] Roll fuchsia [assembly] Deterministically add bootfs packages
Use a PackageSet to ensure that bootfs packages are added in a
determinisitic order no matter which order they are listed in the image
assembly config.

Ok... so here's the problem...

Depending on the iteration order of the bootfs packages, the actual ZBI
may differ. Not functionally. Just the hash of the ZBI.

This can happen because the zbi tool attempts to de-duplicate files to
save storage. Instead of using the file content to compare them, it uses
the st_dev/st_ino attributes, which causes de-duping to only happen when
the two source files are links to the same file, rather than separate
files with the same content.

If the order of iteration changes, then a different source path may be
used for the same bootfs destination, which can cause de-duping to occur
or stop occuring, even if the files are all the same.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/880671
Original-Revision: 1decb8402247b210079de758ca689563686270f9
GitOrigin-RevId: 8363fb67ae8a288ff181e46755897c7d2385db94
Change-Id: I579c80eabf3b65a764d834614dd6aec5fae40c5b