[roll] Roll fuchsia [rbe][rust] Expand linker scripts
Remote linking (e.g. for Rust) may encounter link args
that are linker scripts, as opposed to archives/binaries.
Linker scripts need to be evaluated to determine
all of the files that will be needed for the eventual link.

With this, we can un-hardcode some of the sysroot files
listed in fuchsia.py, and make it more robust to sysroot updates.

Original-Bug: 129842
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/879972
Original-Revision: b8e9f96ff0853125d84305dbc37e833cce526baf
GitOrigin-RevId: bcb3795ee7ebd488a703419db4eeb112da72a644
Change-Id: I3f7c982559192eb54a35a6b3b88be1186b1f4069