[roll] Roll fuchsia [bsdsocket_test] Longer ReadAfterShutdown timeout
Increase the receive timeout on the read-after-shutdown from 1s to 2s
because the call to `AssertBlocked` which asserts that the recv call
is blocked but not yet timed out for 50ms can sometimes take so long
to run in infra that the receive timeout is hit.

The increase is conservative because all test cases where the socket
is still readable after shutdown will take the full timeout duration
to complete.

Original-Fixed: 130000
Multiply: netstack2_fastudp_bsdsocket_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/880975
Original-Revision: 989bb958f183756436d3c3186a89557d6f3ecd96
GitOrigin-RevId: 7f8ab81b511d5331699b5466f4a016e3ba8ea0e2
Change-Id: I1185a12d0ae5f0c4a392ad85bb0654adab7ccae4