[roll] Roll fuchsia [assembly] Remove bundles:debugging
This target is moved into assembly for all eng/userdebug builds.

pistress got moved into bootfs instead of being a shell package, so that
it can still be used for kernel development. Previously, it was added to
//bundles:debugging, which in turn was added to bootfs_labels, but
because it was a shell package it was never added to any products. We
add it as a dep to a platform AIB to ensure it does not bit-rot, but we
do not include it in any products by default.

Original-Bug: b/288301936, b/288301471
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/873816
Size-Review: Aidan Wolter <awolter@google.com>
Original-Revision: 1a52f9d98c2683bf2bf6373df4497a0807aa41c1
GitOrigin-RevId: f34fb7b77617f6dc6cc0a1fed415459108c4afe1
Change-Id: I1518aa66813fa98d2bbd0ac2b517806e938cd082