[roll] Roll fuchsia [sparse] Separate sparse to a library
Allows code handling the sparse format to be reused (in particular, in
the paver).

The interface is slightly different from the interface in
firmware/lib/storage, in that there are no GPT details involved in the
interface, and a generic handle type is used for I/O rather than raw
buffers.  This is more flexible and accommodates the paver's use-case,
where the paver will write directly to a block device representing a
single partition in the GPT.

The interface in firmware/lib/storage is now a thin shim over the
sparse library.

Original-Bug: 129533
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/878652
Original-Revision: 7d75628c0f8f90cf405cc1b5b926f7deda0d4897
GitOrigin-RevId: 0608115be49d1189c78845ab2142a7e3bd3f8603
Change-Id: I52511d855f6bdfd56c0528a85915b01e1bd31a7d