[roll] Roll fuchsia Reland "[fuchsia-controller] Handle cloned proxies in parallel."
This is a reland of commit 4b7548f1411968d5ed2c852291422011cff2efba

Updates: adds a queue wrapper that checks if the loop in which the queue
was created still exists, deleting and creating a new queue if that's
not the case (with tests).

This also removes unnecessary checking of whether a reader has been
added to a loop, and if a reader is going to be removed from a loop
(there can only ever be one reader per file descriptor in a loop). If a
loop is destroyed the readers will be removed anyway.

Multiply: fuchsia_controller_e2e_test
Multiply: fuchsia_controller_fidl_client_test

Original change's description:
> [fuchsia-controller] Handle cloned proxies in parallel.
>
> When running protocol methods in parallel, messages often come in out of
> order, or a protocol will receive a txid for a different method than the
> one that has currently received a read.
>
> This fixes this issue by refactoring the client class to track "staged"
> messages that haven't yet been delivered to the appropriate method
> calls by tracking the txid's correctly.
>
> Multiply: fuchsia_controller_e2e_test
> Multiply: fuchsia_controller_fidl_client_test
> Original-Fixed: 129390
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/876861

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/880973
Original-Revision: 5f88033f2d2f7cf61cfa7a16aae4bc99edad9805
GitOrigin-RevId: f9e1072bd338f6a2a91da74e0651339fca6cd9c9
Change-Id: Ib81df1147ce78d3e0b2df231594418d3c9e52563