[roll] Roll fuchsia [inspect] Route InspectSink to fidl
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881635
Original-Revision: 4f8c1bbe50941d3cbab89d04492a54008c5fb989
GitOrigin-RevId: a8f3307581219a54d1dc8d634685d878542b5286
Change-Id: Idb56f8e0ada61c412294b2672205de834e6e9447