[roll] Roll fuchsia [inspect] Route InspectSink in //src/sys
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881711
Original-Revision: e31b4bcfca0bcdacec391ea0cf79a0c75564cec9
GitOrigin-RevId: c7753da94235533e2c3a63980f24fb2129a6e228
Change-Id: Ide430c4c0af1459879249d6b25203d4f7d446df8