[roll] Roll fuchsia [microbenchmarks] Add Vmo/Memcpy/WithPrecommit benchmarks
Adds some microbenchmarks that measure the time it takes to copy various
amounts of data into a VMO, either with or without committing the pages
in the destination prior to copying. This is an indirect way of
measuring the overhead induced by page faulting during a copy.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858406
Original-Revision: 80abc8073b0440a9d2d4bf73a5382c732657b9c7
GitOrigin-RevId: 411068781cfc93bae6b96d32b4549157c28c9ba0
Change-Id: Icaef003f6e3840454f175fb0ab6f049900fdb49e