[roll] Roll fuchsia [inspect] Route InspectSink in performance
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881639
Original-Revision: ca06d0dc12c90cc60d4eafbef14957eb9094b044
GitOrigin-RevId: 0d400d6f03d311b2e628d6275080d100668e11a8
Change-Id: Ica0053b272d3958c753eae6bd57e5ab618b3a8b4