[roll] Roll fuchsia [inspect] Route InspectSink in //src/factory
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881643
Original-Revision: 250189d335cb9b672e120edfa7c05d50de212204
GitOrigin-RevId: 1f3b8a23f0b59e94d0e6e361b03e5c842d35896c
Change-Id: If09dcffb2be1940a504e85eb8f3e0427841c408e