[roll] Roll fuchsia [ui][input pipeline] InspectHandler: add latency measurements
Add a measurement of input latency from a binding to the
inpsect handler.

Note that Scene Manager instantiates two `InspectHandler`s:
one at the head of the input pipeline, and one at the tail.
Taken together, this means that we can infer both: latency
from the binding into the core of the input pipeline, and
latency within the input pipeline.

Further note, that "within the input pipeline" refers to
the the `InputPipeline` struct, not the library.

For example:
* An `InputEvent` could, theoretically, make its way
  through the entire `InputPipeline` before the handler
  for that event responds to a client's FIDL request for
  the data. In this case, the client-perceived latency
  would be higher than reported by `InspectHandler`.
* Alternatively, a handler might provide a FIDL client
  with event data before the event makes its way to
  the end of the pipeline. In this case, the client-
  perceived latency would be lower than reported by
  `InspectHandler`.

Nonetheless, I believe that understanding latency into
the `InputPipeline`, and through the pipeline stages,
will be valuable in debugging.

Design note: the data is captured in an exponential
histogram, where each bucket is 10x as large as the
previous bucket. My thinking here is that, e.g., the
difference between 100 seconds and 101 seconds is less
important than the difference between 1 msec and 10msec.

Original-Bug: 128663
Test: newly added unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868767
Original-Revision: 80ce814b86b05896658707d6343001bc87b26fda
GitOrigin-RevId: 988b9dd7091c13b64619e7b5da5e317071938cd5
Change-Id: Ieab50adc2580331bc8eae0f8331ff50cf103f637