[roll] Roll fuchsia [inspect] Route InspectSink in //src/hwinfo
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881712
Original-Revision: b6635936fd1b718f34923f80dad65fcd252101d2
GitOrigin-RevId: 76cff214d5491970d5f659bd5baabe30733f054a
Change-Id: I19aa71f07bfa681980158a34ecfc5537c7cc2a56