[roll] Roll fuchsia [inspect] Route InspectSink in //src/testing/sl4f
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881753
Original-Revision: c3e963f198045f912d66357e1922ea76e2097a13
GitOrigin-RevId: c17f9521b171c0002a21a689dfb16704147ad6ff
Change-Id: I7ae17654d76865be7cdcc99136f57c8b5684cf07