[roll] Roll fuchsia [inspect] Route InspectSink in //src/fonts
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881751
Original-Revision: d7027e3fb4251c1cd37c8089020ffb9699d93df5
GitOrigin-RevId: 87e6d5dd5225a0b1073e7be7e92f36de67983369
Change-Id: Ibce5043ed47d820b3ea1018253910592385d61c1