[roll] Roll fuchsia [inspect] Route InspectSink in //sdk/lib/fdio
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881756
Original-Revision: 85baab721615eb77cb07f456c5c4ef01648ee1f3
GitOrigin-RevId: 9d2ce4d65fdf46b029e10b4ad98b6bef4087a79f
Change-Id: Ia5ca5d574cea3450438e9ca2bbf14cf2b286a3b7