[roll] Roll fuchsia [inspect] Route InspectSink in intl
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881648
Original-Revision: fbad1e19536e76cd0c2279abebc751acfd1dc5af
GitOrigin-RevId: e83b5802e2fe2916fe0ca5089242dfb2bff0912f
Change-Id: I837d043875e1a80bd0909d56e3936f5113033e00