[roll] Roll fuchsia [assembly] Remove bundles/bringup:manual_testing
All the contents of manual_testing have been moved to assembly.

mutex_pi_exerciser is moved to a bootfs file instead of a shell package
so that it can still be used for kernel development. It was not
previously added because the shell package was added bootfs_labels
(which does not add shell packages).

Original-Bug: b/288304240
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/873797
Size-Review: Aidan Wolter <awolter@google.com>
Original-Revision: 08a69050df66926400ef95881a39632c8bcecb5d
GitOrigin-RevId: 3eeb80d6fa9a7e9786dfa81577f2f6ff83ea4511
Change-Id: I2bc9edc9eb5a35b0ab72dd4fec44d742ef32bdcb