[roll] Roll fuchsia [inspect] Route InspectSink in //src/developer/remote-control
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881765
Original-Revision: 24e277e6b201f78d02126bff654f785f96e24a46
GitOrigin-RevId: 6f86ec374cba6eb18d66c2042a83a4c29c7dadd3
Change-Id: I83dfcf88c5bb71ce6fc270a7c3266d2bd9a1f6ce