[roll] Roll fuchsia [inspect] Route InspectSink in //src/lib/pty
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881770
Original-Revision: e0ee5d466377076a9fdc8ec416ccf0a6a6964869
GitOrigin-RevId: a2eef16ec0164148aee40cfa0669b9b1d611af15
Change-Id: I29994b8b8067d3005d66f77c30126b7fac5dcca0