[roll] Roll fuchsia [inspect] Route InspectSink in //src/lib/fake-clock
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881758
Original-Revision: 9205070b07d0f9261805ed270657449e0c5cfccf
GitOrigin-RevId: 6c036d69f9e81bc267c79df362406f82470dec49
Change-Id: Ifeab9425d46e65d2fcd18dafc47f1538d2879ab9