[roll] Roll fuchsia [inspect] Route InspectSink in //src/cobalt, //src/developer/forensics
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881764
Original-Revision: 8c076d6b23a63634da61e7b86882356bc3fc0ce9
GitOrigin-RevId: 890103d7c0543a183f3ec52e0ce37ca4639e43b2
Change-Id: I4b7a154917522df79dc520eb8a3516ec30224c5a