[roll] Roll fuchsia [vfs] Apply the typestate pattern [1] on node names
fuchsia.io/Name has a lot of constraints, but not all of them are
respected in the VFS. There are inconsistent checks in the `simple`
directory and path parsing as to what constitutes a valid filesystem
name.

This CL introduces a `Name` type, which wraps a String, and can only be
constructed if the string satisfies the conditions for a valid name, as
spelled out in the fuchsia.io spec. This way as long as we see a `Name`
object we know the name must be valid, and does not have to check again.

This CL updates the DirectlyMutable trait to take `Name`. I couldn't
update `TraversalPosition::Name` because of a non-fuchsia.git repository
depending on it.

[1]: http://cliffle.com/blog/rust-typestate/

Original-Bug: 130228

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881213
Original-Revision: f36cdf4aebc6b03e4023e7ba82c4cc76f48825c3
GitOrigin-RevId: 335fad3b3fc865d2e56645661422c052063031ce
Change-Id: I176fc7b1c22ffff6c321021cf8b65019e7b6eb5a