[roll] Roll fuchsia [inspect] Route InspectSink in //src/lib/fuchsia*
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881766
Original-Revision: dc221fcd4ea414d60489b6c8fbaa045dddbcacb6
GitOrigin-RevId: 5781069e4a809af4ea55d59c038c45e0f24f32b1
Change-Id: I4ef2f49f26f9e1b562fd7cdf7cfa94d24c5ee961