[roll] Roll fuchsia [inspect] Route InspectSink in driver_test_realm and //src/devices
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881641
Original-Revision: daa670fbea182521c3af0ad524e9bdcddbeb5658
GitOrigin-RevId: 5f204234520b5b25459d8e808ad157841d6c2856
Change-Id: I67970d3255dcdc96afde5ede48070606e29eadd9