[roll] Roll fuchsia [inspect] Route InspectSink in //src/chromium
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881762
Original-Revision: cc544f888dd3472611646ba122bf5e9ea563f038
GitOrigin-RevId: d02daf706f5b86dee3fe98c0a9e6af640d8a69aa
Change-Id: Ife7756b752f592c871b58027cd5c43ed8d913e62