[roll] Roll fuchsia [inspect] Route InspectSink in //src/bringup
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881649
Original-Revision: 5e98c0ac15726d4dffd4227d31b1593b05732846
GitOrigin-RevId: 979114a338a2735febc2f4b807a7cac60571e713
Change-Id: I60deae9556e95cf1f22840c7f9189de8cb841cd4