[roll] Roll fuchsia [memalloc] Introduce Pool::NormalizeRanges()
This utility provides a callback with a normalized view of all tracked
ranges, reducing and merging types as desired. In mapping in the
entirety of RAM in physboot, it is useful to generate a normalized view
of RAM, removing any distinction between free and allocated RAM: we
update one site to use this method where we hand-roll this logic today.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/880981
Original-Revision: 05b25a3ba1ba0aa41d37cf04dbc98d04b9aa6ad1
GitOrigin-RevId: 2555eefa898f8677013b72c6adafa4db47b360a9
Change-Id: I344940574e8374afc57280559139753824ee277a