[roll] Roll fuchsia [inspect] Route InspectSink in //src/recovery
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881814
Original-Revision: 2dc8d96177a21d96c45b94386791903f169fcb48
GitOrigin-RevId: 79cecf5c03cee2f384d7934b0f7490c3d0fd30a6
Change-Id: Id5e50f6593f9cef184c94a695328314ea59babd0