[roll] Roll fuchsia [netlink] enable/disable links via RTM_SETLINK
Allow NETLINK_ROUTE sockets to enable and disable links by sending
`RTM_SETLINK` requests.

At the moment, only disabling links is synchronized with the events
observed on the interface watcher, as it's not currently possible to
synchronize enable requests. This is because the interface watcher only
reports whether the interface is online, which is a combination of if
it's enabled, and if the underlying link is UP. This will be addressed
down the road if it's determined to be an issue for clients of
NETLINK_ROUTE.

Original-Fixed: b/283136220
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/880052
Original-Revision: 4fbf88168293bbdaf78d06daa55738f450d7988a
GitOrigin-RevId: 57e327aa6132431cc912fd02961a276c9c01c373
Change-Id: I8c5fcdfdc274931b04d714a19a6edfbd1e8578fd