[roll] Roll fuchsia [ffx][fastboot] consider boot timeout a success
When issuing a "boot", "reboot", or "reboot-bootloader", it's
possible that the "okay" response gets lost in a buffer. So
inside of consdering a lack of an "okay" in those cases as a
a timeout failure, consider them successes.

Also, since we are now considering a timeout to be an expected
result, let's timeout in only a few seconds instead of waiting
30 seconds.

This change includes a fix to fastboot::read_with_timeout() to
actually use a timeout, rather than blocking indefinitely
and checking the timestamp after the fact.

Original-Fixed: b/290218326
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882214
Original-Revision: 021ad1fce51d7588f1ad02d21c03bcdec167d618
GitOrigin-RevId: a6aaef4e6118bd2a97507548668403fdbd0ee2b3
Change-Id: Ia70b13fcd5f07d139d20707223c2decf094237d9