[roll] Roll fuchsia [fdf] Fully support specs in ffx driver list-composites
Update ffx driver list-composites so that it contain all
composite node specs, including the ones not yet matched
to a driver, and is supported in DFv1 and DFv2.

The list of CompositeInfo from specs is provided by
CompositeNodeSpecManager.

This CL also refactors ParentSetCollector so that it
contains the composite driver information and is
responsible for assembling the composite node.

Original-Bug: 119947
Tests: driver-runner-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/877417
Original-Revision: f9f6eedc7175c0ea86bcf83fecaeb215e519eef0
GitOrigin-RevId: 584d6f5044c1856f6b6ab20a66e12bdd7b40c7df
Change-Id: I35351ba61482a8b576daa03cfbc41c9ed63a5cc7