[roll] Roll fuchsia [tracing] Add a core shard target
As part of migrating tracing to be fully optional, it is time to remove the
core shard from non eng builds. However, out of tree defined products
still may depend on the core shard so they need a way to depend on it.

We can't add it first because we get duplicate includes.
We can't remove it first because it will temporarily not be defined.

So we set up a migration where first we land this patch.
Then we point the out of tree products to use this empty target.
Then in the following patch, we simultaneous remove the core shard and
add it to this target.

Original-Bug: 112433
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881768
Original-Revision: 6e0bf6abc759b803272d30639e9cc897a716587d
GitOrigin-RevId: 43fcc07cd502e639434ae35bda3a9c6e0c30f5f8
Change-Id: Idc1d6cb4e1597764d5e2db7abdc92b015a857ec2