[roll] Roll fuchsia [inspect] Route InspectSink to component_manager and related
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881471
Original-Revision: 33b0bf37aa5de4be29534e5f4973a253571afbbf
GitOrigin-RevId: 8b548ddc315f7f19c0b9acc659ebfa3e9ce51538
Change-Id: I77650f1827fe187e774ba317fd48e933219d9202