[roll] Roll fuchsia [inspect] Route InspectSink to ctf
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881531
Original-Revision: 24958fb664155fe79f1354cf968c4e8577c2257d
GitOrigin-RevId: 25f310ae40ff6be1b5c9e621b36d637b756c44ce
Change-Id: Id8f5d1dcefa3dc8374951a59236702ff21d6a26a