[roll] Roll fuchsia [inspect] Route InspectSink to testing
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881532
Original-Revision: d41ba337710627e57fb77818bf11ba87d880e3e0
GitOrigin-RevId: a15ed092b8ba1c628144ddc6f101da124b0c0165
Change-Id: Icb54092207d763bb6eae66fb19bde19e856e22df