[roll] Roll fuchsia [driver_runner] Remove recursion in ScheduleStopComponent
We observe the following sequence of calls can occur resulting in
recursion:

-> CheckForRemoval
  -> ScheduleStopComponent
    -> FinishRemoval
      -> RemoveChild
        -> CheckForRemoval

ScheduleStopComponent only eagerly calls FinishRemoval in the case that
the driver component is not there. By posting an async task to call it
later we break the recursion cycle and ensure the same flow of is
observed in both the optimistic case as well as the slow case.

Additionally, a number of clang tidy lints were addressed.

Test: fx test driver-runner-test.cm
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882586
Original-Revision: f8e6f8e3139afd2a369cd333dac7caf7cf2fbad3
GitOrigin-RevId: 56413779b67d5d1ce8834a582b8b37c86f13fc8f
Change-Id: I34d75516ad0ee3279882c5a40916a65d016eeefe