[roll] Roll fuchsia [inspect] Route InspectSink in ui and graphics
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/offer.shard.cml if the
CML represents a final component (ie non-shard).

The code for the automated routing is here: fxr/855245.
It essentially walked the fuchsia.git tree and opened each CML.
Then it checked whether there were children or collections.
If no, it finished and didn't write to that file.
If yes, it checked if the file was a shard.
If yes, collect the names of the children and collections and insert an offer of InspectSink.
If no, check if the file had the inspect/client.shard.cml file.
If yes, done.
If no, insert inspect/offer.shard.cml.

After this was done, I ran fx format-code and ran fx build,
making manual tweaks until cmc and scrutiny were both happy.

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881640
Original-Revision: 88df614bf02fb670dc6c74c5b2d51b142c3b59c3
GitOrigin-RevId: d147f9caaa50ca4a076e58198e4840aa8a40457d
Change-Id: I4c2d132055b276255241777bf8b7cc480236808b