[roll] Roll fuchsia [inspect] Route InspectSink in //src/power
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881759
Original-Revision: de31b64c16a2e97b0214055aaea17e9555f5423b
GitOrigin-RevId: 5ec256d7707f4491848edacceb86bd55a69e69c0
Change-Id: I418014b08edcace3d83d5368700dd322c7d9ce70