[roll] Roll fuchsia [sdmmc][fidl] Add SDMMC metadata FIDL table
Add a FIDL table to hold metadata for the SDMMC core driver. Include the
fields from the eMMC config struct, and add a `removable` field to be
used for SD card slots. This field will determine whether or not probe
failures are considered fatal and, eventually, whether or not to enable
detection of card insertion/removal.

Original-Bug: 130158
Test: CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882169
Original-Revision: 8fe73cdc34ee97f57bcce8c414731981e68e9260
GitOrigin-RevId: 7b712fd0d4792b6771a3600a2ad967e15b8de51e
Change-Id: I430e2b121233587806b7328863bcf76bf8990288