[roll] Roll fuchsia [inspect] Route InspectSink in //src/identity, fuchsia-hyper-test-support
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881816
Original-Revision: 2afa8390b6c37ca5ac69b03426f97109d9143731
GitOrigin-RevId: 0c926bed98f40b6c64b727949a346dabca6eccd1
Change-Id: I1e1abcb8ca930c28cc46b45fd04d70ff0a8711b4