[roll] Roll fuchsia [component][fuzzing] Make realmfuzzer target library optional
This CL makes realmfuzzer's target library treat a failure to connect to
`fuchsia.fuzzing.CoverageDataCollector` as a non-error. If an executable
is linked against the library but run outside the fuzzing environment,
it will be unable to find a destination for its coverage data but should
still operate normally.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/810867
Original-Revision: 67253791789593ba9ede516efe1921f80a630acd
GitOrigin-RevId: 8cdd85a6cae8cc0d181b1692aff650fdb32b3096
Change-Id: Ie0e03cda476619913c7fac0d325158dc71007dc4