[roll] Roll fuchsia [cm][test] Assert we can connect to protocol in ns without program
We'll likely be relying on this functionality on the next evolution of
how to connect to protocols from ffx/tooling. I didn't find a test
exercising this behavior.

I made a separate test given that one of the test cases in
realm_integration_test relies on the echo server component not being
resolved. The new test would cause it to resolve, therefore, when
running these tests in parallel (which is the default) there could be
flakes.

Original-Bug: b/289268551

Multiply: realm_integration_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/880916
Original-Revision: 8712dfdcaac3baae10c80f6ba0b92f144fc7efbf
GitOrigin-RevId: 5b16362412c35e7fd970dcd50088d9172d16a89c
Change-Id: Ie1ebd25876bcc3be790f7405df49510edb90b12d