[roll] Roll fuchsia [ext4] Remove usage of deprecated fuchsia.mem/Buffer.
I'm working on adding ext4 mount support back to Starnix
and the current `Buffer` API is hard to integrate with
Starnix's `FileOps::get_vmo() -> Arc<Vmo>` API, since we'd
need to clone the VMO to be able to put an owned VMO handle
in the Buffer struct.

Because `Buffer` is supposed to be deprecated in favor of
VMO content size, I'm removing it from the ext4 library rather
than trying to work around the API mismatch in Starnix.

Original-Bug: 129705

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882739
Original-Revision: 26d751920ea032695a553d463aae6f3a1cde09ef
GitOrigin-RevId: d7213971606bceff78a91782ebdfb46e70d07fa4
Change-Id: I3ef726268e8aa26b9296bdff7146142673a3a357