[roll] Roll fuchsia [security] Rename "minfs" directory policy to "data"
The "minfs" directory was renamed to "data" in change
Ic958469b022ea86877ddf464937a6544b54ca913

This disabled the allowlist for `data`, so any component that had
a route for the capability could use it without restriction. Renaming
the directory to `data` in the allowlist fixes the enforcement.

* Allow `/core/session-manager` to use `data`, required as of
  change Icd9c40d876cd4ef443abbef32adf64a9052daa99

Original-Bug: 130281
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882164
Original-Revision: 5cb6c4bcd360f7ddee8721eeea264ba62a60bffd
GitOrigin-RevId: dd9ef9d6dffd87ded261adb1f57d466004f0f794
Change-Id: I99e8e6e28947a3ab82d08a9da345dea504d54bf6