[roll] Roll fuchsia [fidl] fixing fidl-fix compilation
This fixes the fidl-fix compilation by adding the version_selection
parameter to the call of fidl::fix::Fix constructor.

This also refactors the processing of the command line so that
tests can be written to exercise the command line processing.

The fidl-fix tests are part of the //tools/fidl/fidlc/tests label.

Original-Bug: 130019
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882892
Original-Revision: d7242de7938db719706816736b6ed6060c6ee8c9
GitOrigin-RevId: 2500040ed7a39250e8a6441d609c67166461a45d
Change-Id: I17dd9f26df815b802ac8fb136e1f4f55712ae0b0