[roll] Roll fuchsia [starnix] Access current task's memory through CurrentTask
This updates most callsites that are accessing memory associated with
the currently executing task to use a MemoryAccessor implementation
directly on CurrentTask instead of accessing the task's MemoryManager.
This is preparation for accessing this memory through mappings directly
in the future. Accesses to non-current tasks will still go through the
MemoryManager.

This also updates the parameter name used for CurrentTask in many
functions to distinguish it more clearly from a Task that may be
associated with a different thread.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881114
Original-Revision: 90dc9a264a9a92730833bb5001f1bf36cf484b0a
GitOrigin-RevId: b127734e7512fc7f1113b2b3f9734200c5d3c816
Change-Id: Ic7e27fc20d41820225ff97eb2cbda1748e9de73a