[roll] Roll fuchsia [kernel] InitInterrupt take enable_interrupt_callback
To match previous implementations order of operations,
InitInterrupt takes a callback that is in charge of unmasking the irq
the device is assigned to.

The main purpose is to match the legacy driver operation sequence
as much as possible. See fxbug.dev/129378 for details on a tracking
bug for revisiting this decision after migration is complete.

Original-Bug: 126785
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/854018
Original-Revision: 3d0b63e2dcd958ca8f92b6479872ac47807fa5c4
GitOrigin-RevId: 01f210549c43aa4deb8c66c4033d3ec0d083859d
Change-Id: Ic28d42a7381f0d1688fad18c328a42062e2177e6