[roll] Roll fuchsia [kernel][vm] Handle zx_process_{read,write}_memory crossing the end of a VMO smaller than its mapping
Zircon allows mapping a VMO into a virtual range which is larger than
the VMO itself. This CL changes the way we handle
process_{read,write}_memory requests that cross the end of the VMO in
such a mapping: instead of returning an error, we now partially fulfill
the request until the end of the VMO.

Trying to operate on a memory range which lies completely past the end
of the VMO still results in a ZX_ERR_OUT_OF_RANGE error.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881941
Original-Revision: 150342f3b798750fd7f4adae82f75e007d57327f
GitOrigin-RevId: 58e55240949fdcb32c69caaf5b34ef7974a6dd60
Change-Id: I8e83f51c131089dc06f793355c7988ece7caba6e