[roll] Roll fuchsia [inspect] Route InspectSink in //src/camera
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881755
Original-Revision: 10d4bb6f05ac30294c8c2fd25f1147738f8096f1
GitOrigin-RevId: 250f028deb698a5ac87c727e97475a3adaa33aef
Change-Id: Id00b414d83e835e9a41dde42ca9f122c4022b895