[roll] Roll fuchsia [Starnix] Add support for test arguments in LTP test runner
Previously all LTP tests were executed as a single binary. This change
allows to specify test binary that's different from the test name and
pass arguments when necessary.
Now the tests list file is interpreted as LTP run files, see
https://github.com/linux-test-project/ltp/blob/master/doc/ltp-run-files.txt .
Each line contains the name of the test as the first word, followed
by the script to run to execute the test. Lines with a single word are
handled the same way as before - i.e. test name is used as the name of
the binary to run.

Also, removed support for test binary enumeration when there is no
test list file.

Original-Bug: b/290795980
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883119
Original-Revision: abbaf88daa2a5f4414e49ced584721da99eace40
GitOrigin-RevId: 23c1e4b2c5435e816e85ab726a35ecf9b494fca3
Change-Id: Ic7ec3e2e6ed5f58c81be67c58eacf0c710f4b4fd