[roll] Roll fuchsia [wlan][policy] Fix add_phy e2e test flake
When the test adds PHY, most of the time it can expect a
`DeviceMonitorRequest::GetSupportedMacRoles` message right away.
Occasionally, however, the call to add PHY may wait to acquire a lock on
PhyManager before any request can be made to DeviceMonitor, causing a
flake because the test expects a `DeviceMonitorRequest` message when
there's none.

This commit adjusts the test expectation to try up to 5 times if no
`DeviceMonitorRequest` is found right away, with a sleep of 1 second
in between.

Original-Fixed: 130436
Test: CQ
Multiply: wlancfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883122
Original-Revision: d5dac39223f43cc1a4d23bf02bd1b647957a0523
GitOrigin-RevId: ce855ce669f7d1718640989fd20331a10c51ae12
Change-Id: I06205e226d7543e93b90effa5e9c01d41708a6e1