[roll] Roll fuchsia [component][fuzzing] Add a realmfuzzer test for standalone executables
This CL adds a test of an executable linked against the realmfuzzer
target library but outside of a fuzzing environment. It verifies that
executable built for fuzz testing can still be used as part of the
broader system.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/810868
Original-Revision: bde9e75d697c5174a362d4095e2ee31901c9cf3a
GitOrigin-RevId: 11230a0ff1aeb4d820c82a97b13b77eb67d55910
Change-Id: I8deae2a87b35c4dc01e611a1f4693edfc2371c11