[roll] Roll fuchsia [riscv64] Fix ABI issue in riscv platform.cc
uint32_t Foo was defined instead of ktl::optional<uint32_t>.
So the value was returned in one register and garbage on the other, the
side effect is that the value is correct but the flag being zero, caused
it to have a valid value and but !value().

Now we address this by adding the proper include and fixing the return
type.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883584
Original-Revision: ff7fb323f4b502d08f8ed2d1a4deae6664182318
GitOrigin-RevId: 66b76bb6cdd1c698739ed801b72954425089b82e
Change-Id: Ib75537f4a4380a2bfe28738a0a212aa8f9e7d431